Scalenie polskich fontów i konwertera znaków do SuplaOled
