+++ src/SuplaOled.h
@@ -107,5 +107,1423 @@ const uint8_t pressure_bits[] PROGMEM = {
     0x8C, 0x81, 0x81, 0x31, 0x08, 0xC0, 0x03, 0x10, 0x08, 0xE0, 0x07, 0x10, 0x00, 0xE0, 0x07, 0x00, 0x00, 0xC0, 0x03, 0x00, 0x00, 0x80,
     0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
 
+const uint8_t ArialMT_Win1250_Plain_10[] PROGMEM = {
+  0x0A, // Width: 10
+  0x0D, // Height: 13
+  0x20, // First Char: 32
+  0xE0, // Numbers of Chars: 224
+
+  // Jump Table
+  0xFF, 0xFF, 0x00, 0x03, // 32:65535
+  0x00, 0x00, 0x04, 0x03, // 33:0
+  0x00, 0x04, 0x05, 0x04, // 34:4
+  0x00, 0x09, 0x09, 0x06, // 35:9
+  0x00, 0x12, 0x0A, 0x06, // 36:18
+  0x00, 0x1C, 0x10, 0x09, // 37:28
+  0x00, 0x2C, 0x0E, 0x07, // 38:44
+  0x00, 0x3A, 0x01, 0x02, // 39:58
+  0x00, 0x3B, 0x06, 0x03, // 40:59
+  0x00, 0x41, 0x06, 0x03, // 41:65
+  0x00, 0x47, 0x05, 0x04, // 42:71
+  0x00, 0x4C, 0x09, 0x06, // 43:76
+  0x00, 0x55, 0x04, 0x03, // 44:85
+  0x00, 0x59, 0x03, 0x03, // 45:89
+  0x00, 0x5C, 0x04, 0x03, // 46:92
+  0x00, 0x60, 0x05, 0x03, // 47:96
+  0x00, 0x65, 0x0A, 0x06, // 48:101
+  0x00, 0x6F, 0x08, 0x06, // 49:111
+  0x00, 0x77, 0x0A, 0x06, // 50:119
+  0x00, 0x81, 0x0A, 0x06, // 51:129
+  0x00, 0x8B, 0x0B, 0x06, // 52:139
+  0x00, 0x96, 0x0A, 0x06, // 53:150
+  0x00, 0xA0, 0x0A, 0x06, // 54:160
+  0x00, 0xAA, 0x09, 0x06, // 55:170
+  0x00, 0xB3, 0x0A, 0x06, // 56:179
+  0x00, 0xBD, 0x0A, 0x06, // 57:189
+  0x00, 0xC7, 0x04, 0x03, // 58:199
+  0x00, 0xCB, 0x04, 0x03, // 59:203
+  0x00, 0xCF, 0x0A, 0x06, // 60:207
+  0x00, 0xD9, 0x09, 0x06, // 61:217
+  0x00, 0xE2, 0x09, 0x06, // 62:226
+  0x00, 0xEB, 0x0B, 0x06, // 63:235
+  0x00, 0xF6, 0x14, 0x0A, // 64:246
+  0x01, 0x0A, 0x0E, 0x07, // 65:266
+  0x01, 0x18, 0x0C, 0x07, // 66:280
+  0x01, 0x24, 0x0C, 0x07, // 67:292
+  0x01, 0x30, 0x0B, 0x07, // 68:304
+  0x01, 0x3B, 0x0C, 0x07, // 69:315
+  0x01, 0x47, 0x09, 0x06, // 70:327
+  0x01, 0x50, 0x0D, 0x08, // 71:336
+  0x01, 0x5D, 0x0C, 0x07, // 72:349
+  0x01, 0x69, 0x04, 0x03, // 73:361
+  0x01, 0x6D, 0x08, 0x05, // 74:365
+  0x01, 0x75, 0x0E, 0x07, // 75:373
+  0x01, 0x83, 0x0C, 0x06, // 76:387
+  0x01, 0x8F, 0x10, 0x08, // 77:399
+  0x01, 0x9F, 0x0C, 0x07, // 78:415
+  0x01, 0xAB, 0x0E, 0x08, // 79:427
+  0x01, 0xB9, 0x0B, 0x07, // 80:441
+  0x01, 0xC4, 0x0E, 0x08, // 81:452
+  0x01, 0xD2, 0x0C, 0x07, // 82:466
+  0x01, 0xDE, 0x0C, 0x07, // 83:478
+  0x01, 0xEA, 0x0B, 0x06, // 84:490
+  0x01, 0xF5, 0x0C, 0x07, // 85:501
+  0x02, 0x01, 0x0D, 0x07, // 86:513
+  0x02, 0x0E, 0x11, 0x09, // 87:526
+  0x02, 0x1F, 0x0E, 0x07, // 88:543
+  0x02, 0x2D, 0x0D, 0x07, // 89:557
+  0x02, 0x3A, 0x0C, 0x06, // 90:570
+  0x02, 0x46, 0x06, 0x03, // 91:582
+  0x02, 0x4C, 0x06, 0x03, // 92:588
+  0x02, 0x52, 0x04, 0x03, // 93:594
+  0x02, 0x56, 0x09, 0x05, // 94:598
+  0x02, 0x5F, 0x0C, 0x06, // 95:607
+  0x02, 0x6B, 0x03, 0x03, // 96:619
+  0x02, 0x6E, 0x0A, 0x06, // 97:622
+  0x02, 0x78, 0x0A, 0x06, // 98:632
+  0x02, 0x82, 0x0A, 0x05, // 99:642
+  0x02, 0x8C, 0x0A, 0x06, // 100:652
+  0x02, 0x96, 0x0A, 0x06, // 101:662
+  0x02, 0xA0, 0x05, 0x03, // 102:672
+  0x02, 0xA5, 0x0A, 0x06, // 103:677
+  0x02, 0xAF, 0x0A, 0x06, // 104:687
+  0x02, 0xB9, 0x04, 0x02, // 105:697
+  0x02, 0xBD, 0x04, 0x02, // 106:701
+  0x02, 0xC1, 0x08, 0x05, // 107:705
+  0x02, 0xC9, 0x04, 0x02, // 108:713
+  0x02, 0xCD, 0x10, 0x08, // 109:717
+  0x02, 0xDD, 0x0A, 0x06, // 110:733
+  0x02, 0xE7, 0x0A, 0x06, // 111:743
+  0x02, 0xF1, 0x0A, 0x06, // 112:753
+  0x02, 0xFB, 0x0A, 0x06, // 113:763
+  0x03, 0x05, 0x05, 0x03, // 114:773
+  0x03, 0x0A, 0x08, 0x05, // 115:778
+  0x03, 0x12, 0x06, 0x03, // 116:786
+  0x03, 0x18, 0x0A, 0x06, // 117:792
+  0x03, 0x22, 0x09, 0x05, // 118:802
+  0x03, 0x2B, 0x0E, 0x07, // 119:811
+  0x03, 0x39, 0x0A, 0x05, // 120:825
+  0x03, 0x43, 0x09, 0x05, // 121:835
+  0x03, 0x4C, 0x0A, 0x05, // 122:844
+  0x03, 0x56, 0x06, 0x03, // 123:854
+  0x03, 0x5C, 0x04, 0x03, // 124:860
+  0x03, 0x60, 0x05, 0x03, // 125:864
+  0x03, 0x65, 0x09, 0x06, // 126:869
+  0xFF, 0xFF, 0x00, 0x00, // 127:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 128:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 129:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 130:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 131:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 132:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 133:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 134:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 135:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 136:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 137:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 138:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 139:65535
+  0x03, 0x6E, 0x0C, 0x07, // 140:878
+  0xFF, 0xFF, 0x00, 0x0A, // 141:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 142:65535
+  0x03, 0x7A, 0x0C, 0x06, // 143:890
+  0xFF, 0xFF, 0x00, 0x0A, // 144:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 145:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 146:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 147:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 148:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 149:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 150:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 151:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 152:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 153:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 154:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 155:65535
+  0x03, 0x86, 0x08, 0x05, // 156:902
+  0xFF, 0xFF, 0x00, 0x0A, // 157:65535
+  0xFF, 0xFF, 0x00, 0x0A, // 158:65535
+  0x03, 0x8E, 0x0A, 0x05, // 159:910
+  0xFF, 0xFF, 0x00, 0x03, // 160:65535
+  0x03, 0x98, 0x04, 0x03, // 161:920
+  0x03, 0x9C, 0x0A, 0x06, // 162:924
+  0x03, 0xA6, 0x0C, 0x06, // 163:934
+  0x03, 0xB2, 0x0A, 0x06, // 164:946
+  0x03, 0xBC, 0x0E, 0x07, // 165:956
+  0x03, 0xCA, 0x04, 0x03, // 166:970
+  0x03, 0xCE, 0x0A, 0x06, // 167:974
+  0x03, 0xD8, 0x05, 0x03, // 168:984
+  0x03, 0xDD, 0x0D, 0x07, // 169:989
+  0x03, 0xEA, 0x07, 0x04, // 170:1002
+  0x03, 0xF1, 0x0A, 0x06, // 171:1009
+  0x03, 0xFB, 0x09, 0x06, // 172:1019
+  0x04, 0x04, 0x03, 0x03, // 173:1028
+  0x04, 0x07, 0x0D, 0x07, // 174:1031
+  0x04, 0x14, 0x0C, 0x06, // 175:1044
+  0x04, 0x20, 0x07, 0x04, // 176:1056
+  0x04, 0x27, 0x0A, 0x05, // 177:1063
+  0x04, 0x31, 0x05, 0x03, // 178:1073
+  0x04, 0x36, 0x05, 0x03, // 179:1078
+  0x04, 0x3B, 0x05, 0x03, // 180:1083
+  0x04, 0x40, 0x0A, 0x06, // 181:1088
+  0x04, 0x4A, 0x09, 0x05, // 182:1098
+  0x04, 0x53, 0x03, 0x03, // 183:1107
+  0x04, 0x56, 0x06, 0x03, // 184:1110
+  0x04, 0x5C, 0x0A, 0x06, // 185:1116
+  0x04, 0x66, 0x07, 0x04, // 186:1126
+  0x04, 0x6D, 0x0A, 0x06, // 187:1133
+  0x04, 0x77, 0x10, 0x08, // 188:1143
+  0x04, 0x87, 0x10, 0x08, // 189:1159
+  0x04, 0x97, 0x10, 0x08, // 190:1175
+  0x04, 0xA7, 0x0A, 0x05, // 191:1191
+  0x04, 0xB1, 0x0E, 0x07, // 192:1201
+  0x04, 0xBF, 0x0E, 0x07, // 193:1215
+  0x04, 0xCD, 0x0E, 0x07, // 194:1229
+  0x04, 0xDB, 0x0E, 0x07, // 195:1243
+  0x04, 0xE9, 0x0E, 0x07, // 196:1257
+  0x04, 0xF7, 0x0E, 0x07, // 197:1271
+  0x05, 0x05, 0x0C, 0x07, // 198:1285
+  0x05, 0x11, 0x0C, 0x07, // 199:1297
+  0x05, 0x1D, 0x0C, 0x07, // 200:1309
+  0x05, 0x29, 0x0C, 0x07, // 201:1321
+  0x05, 0x35, 0x0C, 0x07, // 202:1333
+  0x05, 0x41, 0x0C, 0x07, // 203:1345
+  0x05, 0x4D, 0x05, 0x03, // 204:1357
+  0x05, 0x52, 0x04, 0x03, // 205:1362
+  0x05, 0x56, 0x04, 0x03, // 206:1366
+  0x05, 0x5A, 0x05, 0x03, // 207:1370
+  0x05, 0x5F, 0x0B, 0x07, // 208:1375
+  0x05, 0x6A, 0x0C, 0x07, // 209:1386
+  0x05, 0x76, 0x0E, 0x08, // 210:1398
+  0x05, 0x84, 0x0E, 0x08, // 211:1412
+  0x05, 0x92, 0x0E, 0x08, // 212:1426
+  0x05, 0xA0, 0x0E, 0x08, // 213:1440
+  0x05, 0xAE, 0x0E, 0x08, // 214:1454
+  0x05, 0xBC, 0x0A, 0x06, // 215:1468
+  0x05, 0xC6, 0x0D, 0x08, // 216:1478
+  0x05, 0xD3, 0x0C, 0x07, // 217:1491
+  0x05, 0xDF, 0x0C, 0x07, // 218:1503
+  0x05, 0xEB, 0x0C, 0x07, // 219:1515
+  0x05, 0xF7, 0x0C, 0x07, // 220:1527
+  0x06, 0x03, 0x0D, 0x07, // 221:1539
+  0x06, 0x10, 0x0B, 0x07, // 222:1552
+  0x06, 0x1B, 0x0C, 0x06, // 223:1563
+  0x06, 0x27, 0x0A, 0x06, // 224:1575
+  0x06, 0x31, 0x0A, 0x06, // 225:1585
+  0x06, 0x3B, 0x0A, 0x06, // 226:1595
+  0x06, 0x45, 0x0A, 0x06, // 227:1605
+  0x06, 0x4F, 0x0A, 0x06, // 228:1615
+  0x06, 0x59, 0x0A, 0x06, // 229:1625
+  0x06, 0x63, 0x0A, 0x05, // 230:1635
+  0x06, 0x6D, 0x0A, 0x05, // 231:1645
+  0x06, 0x77, 0x0A, 0x06, // 232:1655
+  0x06, 0x81, 0x0A, 0x06, // 233:1665
+  0x06, 0x8B, 0x0A, 0x06, // 234:1675
+  0x06, 0x95, 0x0A, 0x06, // 235:1685
+  0x06, 0x9F, 0x05, 0x03, // 236:1695
+  0x06, 0xA4, 0x04, 0x03, // 237:1700
+  0x06, 0xA8, 0x05, 0x03, // 238:1704
+  0x06, 0xAD, 0x05, 0x03, // 239:1709
+  0x06, 0xB2, 0x0A, 0x06, // 240:1714
+  0x06, 0xBC, 0x0A, 0x06, // 241:1724
+  0x06, 0xC6, 0x0A, 0x06, // 242:1734
+  0x06, 0xD0, 0x0A, 0x06, // 243:1744
+  0x06, 0xDA, 0x0A, 0x06, // 244:1754
+  0x06, 0xE4, 0x0A, 0x06, // 245:1764
+  0x06, 0xEE, 0x0A, 0x06, // 246:1774
+  0x06, 0xF8, 0x09, 0x05, // 247:1784
+  0x07, 0x01, 0x0A, 0x06, // 248:1793
+  0x07, 0x0B, 0x0A, 0x06, // 249:1803
+  0x07, 0x15, 0x0A, 0x06, // 250:1813
+  0x07, 0x1F, 0x0A, 0x06, // 251:1823
+  0x07, 0x29, 0x0A, 0x06, // 252:1833
+  0x07, 0x33, 0x09, 0x05, // 253:1843
+  0x07, 0x3C, 0x0A, 0x06, // 254:1852
+  0x07, 0x46, 0x09, 0x05, // 255:1862
+
+  // Font Data
+  0x00,0x00,0xF8,0x02, // 33
+  0x38,0x00,0x00,0x00,0x38, // 34
+  0xA0,0x03,0xE0,0x00,0xB8,0x03,0xE0,0x00,0xB8, // 35
+  0x30,0x01,0x28,0x02,0xF8,0x07,0x48,0x02,0x90,0x01, // 36
+  0x00,0x00,0x30,0x00,0x48,0x00,0x30,0x03,0xC0,0x00,0xB0,0x01,0x48,0x02,0x80,0x01, // 37
+  0x80,0x01,0x50,0x02,0x68,0x02,0xA8,0x02,0x18,0x01,0x80,0x03,0x80,0x02, // 38
+  0x38, // 39
+  0xE0,0x03,0x10,0x04,0x08,0x08, // 40
+  0x08,0x08,0x10,0x04,0xE0,0x03, // 41
+  0x28,0x00,0x18,0x00,0x28, // 42
+  0x40,0x00,0x40,0x00,0xF0,0x01,0x40,0x00,0x40, // 43
+  0x00,0x00,0x00,0x06, // 44
+  0x80,0x00,0x80, // 45
+  0x00,0x00,0x00,0x02, // 46
+  0x00,0x03,0xE0,0x00,0x18, // 47
+  0xF0,0x01,0x08,0x02,0x08,0x02,0x08,0x02,0xF0,0x01, // 48
+  0x00,0x00,0x20,0x00,0x10,0x00,0xF8,0x03, // 49
+  0x10,0x02,0x08,0x03,0x88,0x02,0x48,0x02,0x30,0x02, // 50
+  0x10,0x01,0x08,0x02,0x48,0x02,0x48,0x02,0xB0,0x01, // 51
+  0xC0,0x00,0xA0,0x00,0x90,0x00,0x88,0x00,0xF8,0x03,0x80, // 52
+  0x60,0x01,0x38,0x02,0x28,0x02,0x28,0x02,0xC8,0x01, // 53
+  0xF0,0x01,0x28,0x02,0x28,0x02,0x28,0x02,0xD0,0x01, // 54
+  0x08,0x00,0x08,0x03,0xC8,0x00,0x38,0x00,0x08, // 55
+  0xB0,0x01,0x48,0x02,0x48,0x02,0x48,0x02,0xB0,0x01, // 56
+  0x70,0x01,0x88,0x02,0x88,0x02,0x88,0x02,0xF0,0x01, // 57
+  0x00,0x00,0x20,0x02, // 58
+  0x00,0x00,0x20,0x06, // 59
+  0x00,0x00,0x40,0x00,0xA0,0x00,0xA0,0x00,0x10,0x01, // 60
+  0xA0,0x00,0xA0,0x00,0xA0,0x00,0xA0,0x00,0xA0, // 61
+  0x00,0x00,0x10,0x01,0xA0,0x00,0xA0,0x00,0x40, // 62
+  0x10,0x00,0x08,0x00,0x08,0x00,0xC8,0x02,0x48,0x00,0x30, // 63
+  0x00,0x00,0xC0,0x03,0x30,0x04,0xD0,0x09,0x28,0x0A,0x28,0x0A,0xC8,0x0B,0x68,0x0A,0x10,0x05,0xE0,0x04, // 64
+  0x00,0x02,0xC0,0x01,0xB0,0x00,0x88,0x00,0xB0,0x00,0xC0,0x01,0x00,0x02, // 65
+  0x00,0x00,0xF8,0x03,0x48,0x02,0x48,0x02,0x48,0x02,0xF0,0x01, // 66
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x08,0x02,0x08,0x02,0x10,0x01, // 67
+  0x00,0x00,0xF8,0x03,0x08,0x02,0x08,0x02,0x10,0x01,0xE0, // 68
+  0x00,0x00,0xF8,0x03,0x48,0x02,0x48,0x02,0x48,0x02,0x48,0x02, // 69
+  0x00,0x00,0xF8,0x03,0x48,0x00,0x48,0x00,0x08, // 70
+  0x00,0x00,0xE0,0x00,0x10,0x01,0x08,0x02,0x48,0x02,0x50,0x01,0xC0, // 71
+  0x00,0x00,0xF8,0x03,0x40,0x00,0x40,0x00,0x40,0x00,0xF8,0x03, // 72
+  0x00,0x00,0xF8,0x03, // 73
+  0x00,0x03,0x00,0x02,0x00,0x02,0xF8,0x01, // 74
+  0x00,0x00,0xF8,0x03,0x80,0x00,0x60,0x00,0x90,0x00,0x08,0x01,0x00,0x02, // 75
+  0x00,0x00,0xF8,0x03,0x00,0x02,0x00,0x02,0x00,0x02,0x00,0x02, // 76
+  0x00,0x00,0xF8,0x03,0x30,0x00,0xC0,0x01,0x00,0x02,0xC0,0x01,0x30,0x00,0xF8,0x03, // 77
+  0x00,0x00,0xF8,0x03,0x30,0x00,0x40,0x00,0x80,0x01,0xF8,0x03, // 78
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x08,0x02,0x08,0x02,0x08,0x02,0xF0,0x01, // 79
+  0x00,0x00,0xF8,0x03,0x48,0x00,0x48,0x00,0x48,0x00,0x30, // 80
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x08,0x02,0x08,0x03,0x08,0x03,0xF0,0x02, // 81
+  0x00,0x00,0xF8,0x03,0x48,0x00,0x48,0x00,0xC8,0x00,0x30,0x03, // 82
+  0x00,0x00,0x30,0x01,0x48,0x02,0x48,0x02,0x48,0x02,0x90,0x01, // 83
+  0x00,0x00,0x08,0x00,0x08,0x00,0xF8,0x03,0x08,0x00,0x08, // 84
+  0x00,0x00,0xF8,0x01,0x00,0x02,0x00,0x02,0x00,0x02,0xF8,0x01, // 85
+  0x08,0x00,0x70,0x00,0x80,0x01,0x00,0x02,0x80,0x01,0x70,0x00,0x08, // 86
+  0x18,0x00,0xE0,0x01,0x00,0x02,0xF0,0x01,0x08,0x00,0xF0,0x01,0x00,0x02,0xE0,0x01,0x18, // 87
+  0x00,0x02,0x08,0x01,0x90,0x00,0x60,0x00,0x90,0x00,0x08,0x01,0x00,0x02, // 88
+  0x08,0x00,0x10,0x00,0x20,0x00,0xC0,0x03,0x20,0x00,0x10,0x00,0x08, // 89
+  0x08,0x03,0x88,0x03,0xC8,0x02,0x68,0x02,0x38,0x02,0x18,0x02, // 90
+  0x00,0x00,0xF8,0x0F,0x08,0x08, // 91
+  0x18,0x00,0xE0,0x00,0x00,0x03, // 92
+  0x08,0x08,0xF8,0x0F, // 93
+  0x40,0x00,0x30,0x00,0x08,0x00,0x30,0x00,0x40, // 94
+  0x00,0x08,0x00,0x08,0x00,0x08,0x00,0x08,0x00,0x08,0x00,0x08, // 95
+  0x08,0x00,0x10, // 96
+  0x00,0x00,0x00,0x03,0xA0,0x02,0xA0,0x02,0xE0,0x03, // 97
+  0x00,0x00,0xF8,0x03,0x20,0x02,0x20,0x02,0xC0,0x01, // 98
+  0x00,0x00,0xC0,0x01,0x20,0x02,0x20,0x02,0x40,0x01, // 99
+  0x00,0x00,0xC0,0x01,0x20,0x02,0x20,0x02,0xF8,0x03, // 100
+  0x00,0x00,0xC0,0x01,0xA0,0x02,0xA0,0x02,0xC0,0x02, // 101
+  0x20,0x00,0xF0,0x03,0x28, // 102
+  0x00,0x00,0xC0,0x05,0x20,0x0A,0x20,0x0A,0xE0,0x07, // 103
+  0x00,0x00,0xF8,0x03,0x20,0x00,0x20,0x00,0xC0,0x03, // 104
+  0x00,0x00,0xE8,0x03, // 105
+  0x00,0x08,0xE8,0x07, // 106
+  0xF8,0x03,0x80,0x00,0xC0,0x01,0x20,0x02, // 107
+  0x00,0x00,0xF8,0x03, // 108
+  0x00,0x00,0xE0,0x03,0x20,0x00,0x20,0x00,0xE0,0x03,0x20,0x00,0x20,0x00,0xC0,0x03, // 109
+  0x00,0x00,0xE0,0x03,0x20,0x00,0x20,0x00,0xC0,0x03, // 110
+  0x00,0x00,0xC0,0x01,0x20,0x02,0x20,0x02,0xC0,0x01, // 111
+  0x00,0x00,0xE0,0x0F,0x20,0x02,0x20,0x02,0xC0,0x01, // 112
+  0x00,0x00,0xC0,0x01,0x20,0x02,0x20,0x02,0xE0,0x0F, // 113
+  0x00,0x00,0xE0,0x03,0x20, // 114
+  0x40,0x02,0xA0,0x02,0xA0,0x02,0x20,0x01, // 115
+  0x20,0x00,0xF8,0x03,0x20,0x02, // 116
+  0x00,0x00,0xE0,0x01,0x00,0x02,0x00,0x02,0xE0,0x03, // 117
+  0x20,0x00,0xC0,0x01,0x00,0x02,0xC0,0x01,0x20, // 118
+  0xE0,0x01,0x00,0x02,0xC0,0x01,0x20,0x00,0xC0,0x01,0x00,0x02,0xE0,0x01, // 119
+  0x20,0x02,0x40,0x01,0x80,0x00,0x40,0x01,0x20,0x02, // 120
+  0x20,0x00,0xC0,0x09,0x00,0x06,0xC0,0x01,0x20, // 121
+  0x20,0x02,0x20,0x03,0xA0,0x02,0x60,0x02,0x20,0x02, // 122
+  0x80,0x00,0x78,0x0F,0x08,0x08, // 123
+  0x00,0x00,0xF8,0x0F, // 124
+  0x08,0x08,0x78,0x0F,0x80, // 125
+  0xC0,0x00,0x40,0x00,0xC0,0x00,0x80,0x00,0xC0, // 126
+  0x00,0x00,0x30,0x01,0x48,0x02,0x4A,0x02,0x49,0x02,0x90,0x01, // 140
+  0x08,0x03,0x88,0x03,0xC8,0x02,0x6A,0x02,0x39,0x02,0x18,0x02, // 143
+  0x40,0x02,0xA8,0x02,0xA4,0x02,0x20,0x01, // 156
+  0x20,0x02,0x20,0x03,0xA8,0x02,0x64,0x02,0x20,0x02, // 159
+  0x00,0x00,0xA0,0x0F, // 161
+  0x00,0x00,0xC0,0x01,0xA0,0x0F,0x78,0x02,0x40,0x01, // 162
+  0x00,0x00,0xF8,0x03,0x40,0x02,0x20,0x02,0x10,0x02,0x00,0x02, // 163
+  0x00,0x00,0xE0,0x01,0x20,0x01,0x20,0x01,0xE0,0x01, // 164
+  0x00,0x02,0xC0,0x01,0xB0,0x00,0x88,0x00,0xB0,0x14,0xC0,0x19,0x00,0x02, // 165
+  0x00,0x00,0x38,0x0F, // 166
+  0xD0,0x04,0x28,0x09,0x48,0x09,0x48,0x0A,0x90,0x05, // 167
+  0x08,0x00,0x00,0x00,0x08, // 168
+  0xE0,0x00,0x10,0x01,0x48,0x02,0xA8,0x02,0xA8,0x02,0x10,0x01,0xE0, // 169
+  0x68,0x00,0x68,0x00,0x68,0x00,0x78, // 170
+  0x00,0x00,0x80,0x01,0x40,0x02,0x80,0x01,0x40,0x02, // 171
+  0x20,0x00,0x20,0x00,0x20,0x00,0x20,0x00,0xE0, // 172
+  0x80,0x00,0x80, // 173
+  0xE0,0x00,0x10,0x01,0xE8,0x02,0x68,0x02,0xC8,0x02,0x10,0x01,0xE0, // 174
+  0x08,0x03,0x88,0x03,0xCA,0x02,0x6A,0x02,0x38,0x02,0x18,0x02, // 175
+  0x00,0x00,0x38,0x00,0x28,0x00,0x38, // 176
+  0x40,0x02,0x40,0x02,0xF0,0x03,0x40,0x02,0x40,0x02, // 177
+  0x48,0x00,0x68,0x00,0x58, // 178
+  0xC0,0x00,0xF8,0x03,0x30, // 179
+  0x00,0x00,0x10,0x00,0x08, // 180
+  0x00,0x00,0xE0,0x0F,0x00,0x02,0x00,0x02,0xE0,0x03, // 181
+  0x70,0x00,0xF8,0x0F,0x08,0x00,0xF8,0x0F,0x08, // 182
+  0x00,0x00,0x40, // 183
+  0x00,0x00,0x00,0x14,0x00,0x18, // 184
+  0x00,0x00,0x00,0x03,0xA0,0x02,0xA0,0x16,0xE0,0x1B, // 185
+  0x30,0x00,0x48,0x00,0x48,0x00,0x30, // 186
+  0x00,0x00,0x40,0x02,0x80,0x01,0x40,0x02,0x80,0x01, // 187
+  0x00,0x00,0x10,0x02,0x78,0x01,0xC0,0x00,0x20,0x01,0x90,0x01,0xC8,0x03,0x00,0x01, // 188
+  0x00,0x00,0x10,0x02,0x78,0x01,0x80,0x00,0x60,0x00,0x50,0x02,0x48,0x03,0xC0,0x02, // 189
+  0x48,0x00,0x58,0x00,0x68,0x03,0x80,0x00,0x60,0x01,0x90,0x01,0xC8,0x03,0x00,0x01, // 190
+  0x20,0x02,0x20,0x03,0xA8,0x02,0x68,0x02,0x20,0x02, // 191
+  0x00,0x02,0xC0,0x01,0xB0,0x00,0x89,0x00,0xB2,0x00,0xC0,0x01,0x00,0x02, // 192
+  0x00,0x02,0xC0,0x01,0xB0,0x00,0x8A,0x00,0xB1,0x00,0xC0,0x01,0x00,0x02, // 193
+  0x00,0x02,0xC0,0x01,0xB2,0x00,0x89,0x00,0xB2,0x00,0xC0,0x01,0x00,0x02, // 194
+  0x00,0x02,0xC2,0x01,0xB1,0x00,0x8A,0x00,0xB1,0x00,0xC0,0x01,0x00,0x02, // 195
+  0x00,0x02,0xC0,0x01,0xB2,0x00,0x88,0x00,0xB2,0x00,0xC0,0x01,0x00,0x02, // 196
+  0x00,0x02,0xC0,0x01,0xBE,0x00,0x8A,0x00,0xBE,0x00,0xC0,0x01,0x00,0x02, // 197
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x0A,0x02,0x09,0x02,0x10,0x01, // 198
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x08,0x16,0x08,0x1A,0x10,0x01, // 199
+  0x00,0x00,0xF8,0x03,0x49,0x02,0x4A,0x02,0x48,0x02,0x48,0x02, // 200
+  0x00,0x00,0xF8,0x03,0x48,0x02,0x4A,0x02,0x49,0x02,0x48,0x02, // 201
+  0x00,0x00,0xF8,0x03,0x48,0x02,0x48,0x16,0x48,0x1A,0x48,0x02, // 202
+  0x00,0x00,0xF8,0x03,0x4A,0x02,0x48,0x02,0x4A,0x02,0x48,0x02, // 203
+  0x00,0x00,0xF9,0x03,0x02, // 204
+  0x02,0x00,0xF9,0x03, // 205
+  0x01,0x00,0xFA,0x03, // 206
+  0x02,0x00,0xF8,0x03,0x02, // 207
+  0x40,0x00,0xF8,0x03,0x48,0x02,0x48,0x02,0x10,0x01,0xE0, // 208
+  0x00,0x00,0xF8,0x03,0x30,0x00,0x42,0x00,0x81,0x01,0xF8,0x03, // 209
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x09,0x02,0x0A,0x02,0x08,0x02,0xF0,0x01, // 210
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x0A,0x02,0x09,0x02,0x08,0x02,0xF0,0x01, // 211
+  0x00,0x00,0xF0,0x01,0x08,0x02,0x0A,0x02,0x09,0x02,0x0A,0x02,0xF0,0x01, // 212
+  0x00,0x00,0xF0,0x01,0x0A,0x02,0x09,0x02,0x0A,0x02,0x09,0x02,0xF0,0x01, // 213
+  0x00,0x00,0xF0,0x01,0x0A,0x02,0x08,0x02,0x0A,0x02,0x08,0x02,0xF0,0x01, // 214
+  0x10,0x01,0xA0,0x00,0xE0,0x00,0xA0,0x00,0x10,0x01, // 215
+  0x00,0x00,0xF0,0x02,0x08,0x03,0xC8,0x02,0x28,0x02,0x18,0x03,0xE8, // 216
+  0x00,0x00,0xF8,0x01,0x01,0x02,0x02,0x02,0x00,0x02,0xF8,0x01, // 217
+  0x00,0x00,0xF8,0x01,0x02,0x02,0x01,0x02,0x00,0x02,0xF8,0x01, // 218
+  0x00,0x00,0xF8,0x01,0x02,0x02,0x01,0x02,0x02,0x02,0xF8,0x01, // 219
+  0x00,0x00,0xF8,0x01,0x02,0x02,0x00,0x02,0x02,0x02,0xF8,0x01, // 220
+  0x08,0x00,0x10,0x00,0x20,0x00,0xC2,0x03,0x21,0x00,0x10,0x00,0x08, // 221
+  0x00,0x00,0xF8,0x03,0x10,0x01,0x10,0x01,0x10,0x01,0xE0, // 222
+  0x00,0x00,0xF0,0x03,0x08,0x01,0x48,0x02,0xB0,0x02,0x80,0x01, // 223
+  0x00,0x00,0x00,0x03,0xA4,0x02,0xA8,0x02,0xE0,0x03, // 224
+  0x00,0x00,0x00,0x03,0xA8,0x02,0xA4,0x02,0xE0,0x03, // 225
+  0x00,0x00,0x00,0x03,0xA8,0x02,0xA4,0x02,0xE8,0x03, // 226
+  0x00,0x00,0x08,0x03,0xA4,0x02,0xA8,0x02,0xE4,0x03, // 227
+  0x00,0x00,0x00,0x03,0xA8,0x02,0xA0,0x02,0xE8,0x03, // 228
+  0x00,0x00,0x00,0x03,0xAE,0x02,0xAA,0x02,0xEE,0x03, // 229
+  0x00,0x00,0xC0,0x01,0x28,0x02,0x24,0x02,0x40,0x01, // 230
+  0x00,0x00,0xC0,0x01,0x20,0x16,0x20,0x1A,0x40,0x01, // 231
+  0x00,0x00,0xC0,0x01,0xA4,0x02,0xA8,0x02,0xC0,0x02, // 232
+  0x00,0x00,0xC0,0x01,0xA8,0x02,0xA4,0x02,0xC0,0x02, // 233
+  0x00,0x00,0xC0,0x01,0xA0,0x02,0xA0,0x16,0xC0,0x1A, // 234
+  0x00,0x00,0xC0,0x01,0xA8,0x02,0xA0,0x02,0xC8,0x02, // 235
+  0x00,0x00,0xE4,0x03,0x08, // 236
+  0x08,0x00,0xE4,0x03, // 237
+  0x08,0x00,0xE4,0x03,0x08, // 238
+  0x08,0x00,0xE0,0x03,0x08, // 239
+  0x00,0x00,0xC0,0x01,0x28,0x02,0x38,0x02,0xE0,0x01, // 240
+  0x00,0x00,0xE0,0x03,0x28,0x00,0x24,0x00,0xC0,0x03, // 241
+  0x00,0x00,0xC0,0x01,0x24,0x02,0x28,0x02,0xC0,0x01, // 242
+  0x00,0x00,0xC0,0x01,0x28,0x02,0x24,0x02,0xC0,0x01, // 243
+  0x00,0x00,0xC0,0x01,0x28,0x02,0x24,0x02,0xC8,0x01, // 244
+  0x00,0x00,0xC8,0x01,0x24,0x02,0x28,0x02,0xC4,0x01, // 245
+  0x00,0x00,0xC0,0x01,0x28,0x02,0x20,0x02,0xC8,0x01, // 246
+  0x40,0x00,0x40,0x00,0x50,0x01,0x40,0x00,0x40, // 247
+  0x00,0x00,0xC0,0x02,0xA0,0x03,0x60,0x02,0xA0,0x01, // 248
+  0x00,0x00,0xE0,0x01,0x04,0x02,0x08,0x02,0xE0,0x03, // 249
+  0x00,0x00,0xE0,0x01,0x08,0x02,0x04,0x02,0xE0,0x03, // 250
+  0x00,0x00,0xE8,0x01,0x04,0x02,0x08,0x02,0xE0,0x03, // 251
+  0x00,0x00,0xE0,0x01,0x08,0x02,0x00,0x02,0xE8,0x03, // 252
+  0x20,0x00,0xC0,0x09,0x08,0x06,0xC4,0x01,0x20, // 253
+  0x00,0x00,0xF8,0x0F,0x20,0x02,0x20,0x02,0xC0,0x01, // 254
+  0x20,0x00,0xC8,0x09,0x00,0x06,0xC8,0x01,0x20 // 255
+};
+
+const uint8_t ArialMT_Win1250_Plain_16[] PROGMEM = {
+  0x10, // Width: 16
+  0x13, // Height: 19
+  0x20, // First Char: 32
+  0xE0, // Numbers of Chars: 224
+
+  // Jump Table
+  0xFF, 0xFF, 0x00, 0x04, // 32:65535
+  0x00, 0x00, 0x08, 0x04, // 33:0
+  0x00, 0x08, 0x0D, 0x06, // 34:8
+  0x00, 0x15, 0x1A, 0x09, // 35:21
+  0x00, 0x2F, 0x17, 0x09, // 36:47
+  0x00, 0x46, 0x26, 0x0E, // 37:70
+  0x00, 0x6C, 0x1D, 0x0B, // 38:108
+  0x00, 0x89, 0x04, 0x03, // 39:137
+  0x00, 0x8D, 0x0C, 0x05, // 40:141
+  0x00, 0x99, 0x0B, 0x05, // 41:153
+  0x00, 0xA4, 0x0D, 0x06, // 42:164
+  0x00, 0xB1, 0x17, 0x09, // 43:177
+  0x00, 0xC8, 0x09, 0x04, // 44:200
+  0x00, 0xD1, 0x0B, 0x05, // 45:209
+  0x00, 0xDC, 0x08, 0x04, // 46:220
+  0x00, 0xE4, 0x0A, 0x04, // 47:228
+  0x00, 0xEE, 0x17, 0x09, // 48:238
+  0x01, 0x05, 0x11, 0x09, // 49:261
+  0x01, 0x16, 0x17, 0x09, // 50:278
+  0x01, 0x2D, 0x17, 0x09, // 51:301
+  0x01, 0x44, 0x17, 0x09, // 52:324
+  0x01, 0x5B, 0x17, 0x09, // 53:347
+  0x01, 0x72, 0x17, 0x09, // 54:370
+  0x01, 0x89, 0x16, 0x09, // 55:393
+  0x01, 0x9F, 0x17, 0x09, // 56:415
+  0x01, 0xB6, 0x17, 0x09, // 57:438
+  0x01, 0xCD, 0x05, 0x04, // 58:461
+  0x01, 0xD2, 0x06, 0x04, // 59:466
+  0x01, 0xD8, 0x17, 0x09, // 60:472
+  0x01, 0xEF, 0x17, 0x09, // 61:495
+  0x02, 0x06, 0x17, 0x09, // 62:518
+  0x02, 0x1D, 0x16, 0x09, // 63:541
+  0x02, 0x33, 0x2F, 0x10, // 64:563
+  0x02, 0x62, 0x1D, 0x0B, // 65:610
+  0x02, 0x7F, 0x1D, 0x0B, // 66:639
+  0x02, 0x9C, 0x20, 0x0C, // 67:668
+  0x02, 0xBC, 0x20, 0x0C, // 68:700
+  0x02, 0xDC, 0x1D, 0x0B, // 69:732
+  0x02, 0xF9, 0x19, 0x0A, // 70:761
+  0x03, 0x12, 0x20, 0x0C, // 71:786
+  0x03, 0x32, 0x1D, 0x0C, // 72:818
+  0x03, 0x4F, 0x05, 0x04, // 73:847
+  0x03, 0x54, 0x14, 0x08, // 74:852
+  0x03, 0x68, 0x1D, 0x0B, // 75:872
+  0x03, 0x85, 0x17, 0x09, // 76:901
+  0x03, 0x9C, 0x23, 0x0D, // 77:924
+  0x03, 0xBF, 0x1D, 0x0C, // 78:959
+  0x03, 0xDC, 0x20, 0x0C, // 79:988
+  0x03, 0xFC, 0x1C, 0x0B, // 80:1020
+  0x04, 0x18, 0x20, 0x0C, // 81:1048
+  0x04, 0x38, 0x1D, 0x0C, // 82:1080
+  0x04, 0x55, 0x1D, 0x0B, // 83:1109
+  0x04, 0x72, 0x19, 0x0A, // 84:1138
+  0x04, 0x8B, 0x1D, 0x0C, // 85:1163
+  0x04, 0xA8, 0x1C, 0x0B, // 86:1192
+  0x04, 0xC4, 0x2B, 0x0F, // 87:1220
+  0x04, 0xEF, 0x20, 0x0B, // 88:1263
+  0x05, 0x0F, 0x19, 0x0B, // 89:1295
+  0x05, 0x28, 0x1A, 0x0A, // 90:1320
+  0x05, 0x42, 0x0C, 0x04, // 91:1346
+  0x05, 0x4E, 0x0B, 0x04, // 92:1358
+  0x05, 0x59, 0x09, 0x04, // 93:1369
+  0x05, 0x62, 0x14, 0x08, // 94:1378
+  0x05, 0x76, 0x1B, 0x09, // 95:1398
+  0x05, 0x91, 0x07, 0x05, // 96:1425
+  0x05, 0x98, 0x17, 0x09, // 97:1432
+  0x05, 0xAF, 0x17, 0x09, // 98:1455
+  0x05, 0xC6, 0x14, 0x08, // 99:1478
+  0x05, 0xDA, 0x17, 0x09, // 100:1498
+  0x05, 0xF1, 0x17, 0x09, // 101:1521
+  0x06, 0x08, 0x0A, 0x04, // 102:1544
+  0x06, 0x12, 0x17, 0x09, // 103:1554
+  0x06, 0x29, 0x14, 0x09, // 104:1577
+  0x06, 0x3D, 0x05, 0x04, // 105:1597
+  0x06, 0x42, 0x06, 0x04, // 106:1602
+  0x06, 0x48, 0x17, 0x08, // 107:1608
+  0x06, 0x5F, 0x05, 0x04, // 108:1631
+  0x06, 0x64, 0x23, 0x0D, // 109:1636
+  0x06, 0x87, 0x14, 0x09, // 110:1671
+  0x06, 0x9B, 0x17, 0x09, // 111:1691
+  0x06, 0xB2, 0x17, 0x09, // 112:1714
+  0x06, 0xC9, 0x18, 0x09, // 113:1737
+  0x06, 0xE1, 0x0D, 0x05, // 114:1761
+  0x06, 0xEE, 0x14, 0x08, // 115:1774
+  0x07, 0x02, 0x0B, 0x04, // 116:1794
+  0x07, 0x0D, 0x14, 0x09, // 117:1805
+  0x07, 0x21, 0x13, 0x08, // 118:1825
+  0x07, 0x34, 0x1F, 0x0C, // 119:1844
+  0x07, 0x53, 0x14, 0x08, // 120:1875
+  0x07, 0x67, 0x13, 0x08, // 121:1895
+  0x07, 0x7A, 0x14, 0x08, // 122:1914
+  0x07, 0x8E, 0x0F, 0x05, // 123:1934
+  0x07, 0x9D, 0x06, 0x04, // 124:1949
+  0x07, 0xA3, 0x0E, 0x05, // 125:1955
+  0x07, 0xB1, 0x17, 0x09, // 126:1969
+  0xFF, 0xFF, 0x00, 0x00, // 127:65535
+  0xFF, 0xFF, 0x00, 0x10, // 128:65535
+  0xFF, 0xFF, 0x00, 0x10, // 129:65535
+  0xFF, 0xFF, 0x00, 0x10, // 130:65535
+  0xFF, 0xFF, 0x00, 0x10, // 131:65535
+  0xFF, 0xFF, 0x00, 0x10, // 132:65535
+  0xFF, 0xFF, 0x00, 0x10, // 133:65535
+  0xFF, 0xFF, 0x00, 0x10, // 134:65535
+  0xFF, 0xFF, 0x00, 0x10, // 135:65535
+  0xFF, 0xFF, 0x00, 0x10, // 136:65535
+  0xFF, 0xFF, 0x00, 0x10, // 137:65535
+  0xFF, 0xFF, 0x00, 0x10, // 138:65535
+  0xFF, 0xFF, 0x00, 0x10, // 139:65535
+  0x07, 0xC8, 0x1D, 0x0B, // 140:1992
+  0xFF, 0xFF, 0x00, 0x10, // 141:65535
+  0xFF, 0xFF, 0x00, 0x10, // 142:65535
+  0x07, 0xE5, 0x1A, 0x0A, // 143:2021
+  0xFF, 0xFF, 0x00, 0x10, // 144:65535
+  0xFF, 0xFF, 0x00, 0x10, // 145:65535
+  0xFF, 0xFF, 0x00, 0x10, // 146:65535
+  0xFF, 0xFF, 0x00, 0x10, // 147:65535
+  0xFF, 0xFF, 0x00, 0x10, // 148:65535
+  0xFF, 0xFF, 0x00, 0x10, // 149:65535
+  0xFF, 0xFF, 0x00, 0x10, // 150:65535
+  0xFF, 0xFF, 0x00, 0x10, // 151:65535
+  0xFF, 0xFF, 0x00, 0x10, // 152:65535
+  0xFF, 0xFF, 0x00, 0x10, // 153:65535
+  0xFF, 0xFF, 0x00, 0x10, // 154:65535
+  0xFF, 0xFF, 0x00, 0x10, // 155:65535
+  0x07, 0xFF, 0x14, 0x08, // 156:2047
+  0xFF, 0xFF, 0x00, 0x10, // 157:65535
+  0xFF, 0xFF, 0x00, 0x10, // 158:65535
+  0x08, 0x13, 0x14, 0x08, // 159:2067
+  0xFF, 0xFF, 0x00, 0x04, // 160:65535
+  0x08, 0x27, 0x09, 0x05, // 161:2087
+  0x08, 0x30, 0x17, 0x09, // 162:2096
+  0x08, 0x47, 0x17, 0x09, // 163:2119
+  0x08, 0x5E, 0x14, 0x09, // 164:2142
+  0x08, 0x72, 0x1D, 0x0B, // 165:2162
+  0x08, 0x8F, 0x06, 0x04, // 166:2191
+  0x08, 0x95, 0x17, 0x09, // 167:2197
+  0x08, 0xAC, 0x07, 0x05, // 168:2220
+  0x08, 0xB3, 0x23, 0x0C, // 169:2227
+  0x08, 0xD6, 0x0E, 0x06, // 170:2262
+  0x08, 0xE4, 0x14, 0x09, // 171:2276
+  0x08, 0xF8, 0x17, 0x09, // 172:2296
+  0x09, 0x0F, 0x0B, 0x05, // 173:2319
+  0x09, 0x1A, 0x23, 0x0C, // 174:2330
+  0x09, 0x3D, 0x1A, 0x0A, // 175:2365
+  0x09, 0x57, 0x0D, 0x06, // 176:2391
+  0x09, 0x64, 0x17, 0x09, // 177:2404
+  0x09, 0x7B, 0x0E, 0x05, // 178:2427
+  0x09, 0x89, 0x0A, 0x04, // 179:2441
+  0x09, 0x93, 0x0A, 0x05, // 180:2451
+  0x09, 0x9D, 0x17, 0x09, // 181:2461
+  0x09, 0xB4, 0x19, 0x09, // 182:2484
+  0x09, 0xCD, 0x08, 0x05, // 183:2509
+  0x09, 0xD5, 0x0C, 0x05, // 184:2517
+  0x09, 0xE1, 0x17, 0x09, // 185:2529
+  0x09, 0xF8, 0x0D, 0x06, // 186:2552
+  0x0A, 0x05, 0x17, 0x09, // 187:2565
+  0x0A, 0x1C, 0x26, 0x0D, // 188:2588
+  0x0A, 0x42, 0x26, 0x0D, // 189:2626
+  0x0A, 0x68, 0x26, 0x0D, // 190:2664
+  0x0A, 0x8E, 0x14, 0x08, // 191:2702
+  0x0A, 0xA2, 0x1D, 0x0B, // 192:2722
+  0x0A, 0xBF, 0x1D, 0x0B, // 193:2751
+  0x0A, 0xDC, 0x1D, 0x0B, // 194:2780
+  0x0A, 0xF9, 0x1D, 0x0B, // 195:2809
+  0x0B, 0x16, 0x1D, 0x0B, // 196:2838
+  0x0B, 0x33, 0x1D, 0x0B, // 197:2867
+  0x0B, 0x50, 0x20, 0x0C, // 198:2896
+  0x0B, 0x70, 0x20, 0x0C, // 199:2928
+  0x0B, 0x90, 0x1D, 0x0B, // 200:2960
+  0x0B, 0xAD, 0x1D, 0x0B, // 201:2989
+  0x0B, 0xCA, 0x1D, 0x0B, // 202:3018
+  0x0B, 0xE7, 0x1D, 0x0B, // 203:3047
+  0x0C, 0x04, 0x05, 0x04, // 204:3076
+  0x0C, 0x09, 0x07, 0x04, // 205:3081
+  0x0C, 0x10, 0x0A, 0x04, // 206:3088
+  0x0C, 0x1A, 0x07, 0x04, // 207:3098
+  0x0C, 0x21, 0x20, 0x0C, // 208:3105
+  0x0C, 0x41, 0x1D, 0x0C, // 209:3137
+  0x0C, 0x5E, 0x20, 0x0C, // 210:3166
+  0x0C, 0x7E, 0x20, 0x0C, // 211:3198
+  0x0C, 0x9E, 0x20, 0x0C, // 212:3230
+  0x0C, 0xBE, 0x20, 0x0C, // 213:3262
+  0x0C, 0xDE, 0x20, 0x0C, // 214:3294
+  0x0C, 0xFE, 0x17, 0x09, // 215:3326
+  0x0D, 0x15, 0x20, 0x0C, // 216:3349
+  0x0D, 0x35, 0x1D, 0x0C, // 217:3381
+  0x0D, 0x52, 0x1D, 0x0C, // 218:3410
+  0x0D, 0x6F, 0x1D, 0x0C, // 219:3439
+  0x0D, 0x8C, 0x1D, 0x0C, // 220:3468
+  0x0D, 0xA9, 0x19, 0x0B, // 221:3497
+  0x0D, 0xC2, 0x1D, 0x0B, // 222:3522
+  0x0D, 0xDF, 0x17, 0x0A, // 223:3551
+  0x0D, 0xF6, 0x17, 0x09, // 224:3574
+  0x0E, 0x0D, 0x17, 0x09, // 225:3597
+  0x0E, 0x24, 0x17, 0x09, // 226:3620
+  0x0E, 0x3B, 0x17, 0x09, // 227:3643
+  0x0E, 0x52, 0x17, 0x09, // 228:3666
+  0x0E, 0x69, 0x17, 0x09, // 229:3689
+  0x0E, 0x80, 0x14, 0x08, // 230:3712
+  0x0E, 0x94, 0x14, 0x08, // 231:3732
+  0x0E, 0xA8, 0x17, 0x09, // 232:3752
+  0x0E, 0xBF, 0x17, 0x09, // 233:3775
+  0x0E, 0xD6, 0x17, 0x09, // 234:3798
+  0x0E, 0xED, 0x17, 0x09, // 235:3821
+  0x0F, 0x04, 0x05, 0x04, // 236:3844
+  0x0F, 0x09, 0x07, 0x04, // 237:3849
+  0x0F, 0x10, 0x0A, 0x04, // 238:3856
+  0x0F, 0x1A, 0x07, 0x04, // 239:3866
+  0x0F, 0x21, 0x17, 0x09, // 240:3873
+  0x0F, 0x38, 0x14, 0x09, // 241:3896
+  0x0F, 0x4C, 0x17, 0x09, // 242:3916
+  0x0F, 0x63, 0x17, 0x09, // 243:3939
+  0x0F, 0x7A, 0x17, 0x09, // 244:3962
+  0x0F, 0x91, 0x17, 0x09, // 245:3985
+  0x0F, 0xA8, 0x17, 0x09, // 246:4008
+  0x0F, 0xBF, 0x17, 0x09, // 247:4031
+  0x0F, 0xD6, 0x17, 0x0A, // 248:4054
+  0x0F, 0xED, 0x14, 0x09, // 249:4077
+  0x10, 0x01, 0x14, 0x09, // 250:4097
+  0x10, 0x15, 0x14, 0x09, // 251:4117
+  0x10, 0x29, 0x14, 0x09, // 252:4137
+  0x10, 0x3D, 0x13, 0x08, // 253:4157
+  0x10, 0x50, 0x17, 0x09, // 254:4176
+  0x10, 0x67, 0x13, 0x08, // 255:4199
+
+  // Font Data
+  0x00,0x00,0x00,0x00,0x00,0x00,0xF8,0x5F, // 33
+  0x00,0x00,0x00,0x78,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x78, // 34
+  0x80,0x08,0x00,0x80,0x78,0x00,0xC0,0x0F,0x00,0xB8,0x08,0x00,0x80,0x08,0x00,0x80,0x78,0x00,0xC0,0x0F,0x00,0xB8,0x08,0x00,0x80,0x08, // 35
+  0x00,0x00,0x00,0xE0,0x10,0x00,0x10,0x21,0x00,0x08,0x41,0x00,0xFC,0xFF,0x00,0x08,0x42,0x00,0x10,0x22,0x00,0x20,0x1C, // 36
+  0x00,0x00,0x00,0xF0,0x00,0x00,0x08,0x01,0x00,0x08,0x01,0x00,0x08,0x61,0x00,0xF0,0x18,0x00,0x00,0x06,0x00,0xC0,0x01,0x00,0x30,0x3C,0x00,0x08,0x42,0x00,0x00,0x42,0x00,0x00,0x42,0x00,0x00,0x3C, // 37
+  0x00,0x00,0x00,0x00,0x1C,0x00,0x70,0x22,0x00,0x88,0x41,0x00,0x08,0x43,0x00,0x88,0x44,0x00,0x70,0x28,0x00,0x00,0x10,0x00,0x00,0x28,0x00,0x00,0x44, // 38
+  0x00,0x00,0x00,0x78, // 39
+  0x00,0x00,0x00,0x80,0x3F,0x00,0x70,0xC0,0x01,0x08,0x00,0x02, // 40
+  0x00,0x00,0x00,0x08,0x00,0x02,0x70,0xC0,0x01,0x80,0x3F, // 41
+  0x10,0x00,0x00,0xD0,0x00,0x00,0x38,0x00,0x00,0xD0,0x00,0x00,0x10, // 42
+  0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0xC0,0x1F,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02, // 43
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xC0,0x01, // 44
+  0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08, // 45
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x40, // 46
+  0x00,0x60,0x00,0x00,0x1E,0x00,0xE0,0x01,0x00,0x18, // 47
+  0x00,0x00,0x00,0xE0,0x1F,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0xE0,0x1F, // 48
+  0x00,0x00,0x00,0x00,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0x00,0x10,0x00,0x00,0xF8,0x7F, // 49
+  0x00,0x00,0x00,0x20,0x40,0x00,0x10,0x60,0x00,0x08,0x50,0x00,0x08,0x48,0x00,0x08,0x44,0x00,0x10,0x43,0x00,0xE0,0x40, // 50
+  0x00,0x00,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x88,0x41,0x00,0xF0,0x22,0x00,0x00,0x1C, // 51
+  0x00,0x0C,0x00,0x00,0x0A,0x00,0x00,0x09,0x00,0xC0,0x08,0x00,0x20,0x08,0x00,0x10,0x08,0x00,0xF8,0x7F,0x00,0x00,0x08, // 52
+  0x00,0x00,0x00,0xC0,0x11,0x00,0xB8,0x20,0x00,0x88,0x40,0x00,0x88,0x40,0x00,0x88,0x40,0x00,0x08,0x21,0x00,0x08,0x1E, // 53
+  0x00,0x00,0x00,0xE0,0x1F,0x00,0x10,0x21,0x00,0x88,0x40,0x00,0x88,0x40,0x00,0x88,0x40,0x00,0x10,0x21,0x00,0x20,0x1E, // 54
+  0x00,0x00,0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08,0x78,0x00,0x08,0x07,0x00,0xC8,0x00,0x00,0x28,0x00,0x00,0x18, // 55
+  0x00,0x00,0x00,0x60,0x1C,0x00,0x90,0x22,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x90,0x22,0x00,0x60,0x1C, // 56
+  0x00,0x00,0x00,0xE0,0x11,0x00,0x10,0x22,0x00,0x08,0x44,0x00,0x08,0x44,0x00,0x08,0x44,0x00,0x10,0x22,0x00,0xE0,0x1F, // 57
+  0x00,0x00,0x00,0x40,0x40, // 58
+  0x00,0x00,0x00,0x40,0xC0,0x01, // 59
+  0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x05,0x00,0x00,0x05,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x40,0x10, // 60
+  0x00,0x00,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x80,0x08, // 61
+  0x00,0x00,0x00,0x40,0x10,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x80,0x08,0x00,0x00,0x05,0x00,0x00,0x05,0x00,0x00,0x02, // 62
+  0x00,0x00,0x00,0x60,0x00,0x00,0x10,0x00,0x00,0x08,0x00,0x00,0x08,0x5C,0x00,0x08,0x02,0x00,0x10,0x01,0x00,0xE0, // 63
+  0x00,0x00,0x00,0x00,0x3F,0x00,0xC0,0x40,0x00,0x20,0x80,0x00,0x10,0x1E,0x01,0x10,0x21,0x01,0x88,0x40,0x02,0x48,0x40,0x02,0x48,0x40,0x02,0x48,0x20,0x02,0x88,0x7C,0x02,0xC8,0x43,0x02,0x10,0x40,0x02,0x10,0x20,0x01,0x60,0x10,0x01,0x80,0x8F, // 64
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x70,0x04,0x00,0x08,0x04,0x00,0x70,0x04,0x00,0x80,0x07,0x00,0x00,0x1C,0x00,0x00,0x60, // 65
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x90,0x22,0x00,0x60,0x1C, // 66
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10, // 67
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 68
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x40, // 69
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08, // 70
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x42,0x00,0x08,0x42,0x00,0x10,0x22,0x00,0x20,0x12,0x00,0x00,0x0E, // 71
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0xF8,0x7F, // 72
+  0x00,0x00,0x00,0xF8,0x7F, // 73
+  0x00,0x00,0x00,0x00,0x38,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0xF8,0x3F, // 74
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x00,0x04,0x00,0x00,0x02,0x00,0x00,0x01,0x00,0x80,0x03,0x00,0x40,0x04,0x00,0x20,0x18,0x00,0x10,0x20,0x00,0x08,0x40, // 75
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40, // 76
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x30,0x00,0x00,0xC0,0x00,0x00,0x00,0x03,0x00,0x00,0x1C,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x00,0x03,0x00,0xC0,0x00,0x00,0x30,0x00,0x00,0xF8,0x7F, // 77
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x10,0x00,0x00,0x60,0x00,0x00,0x80,0x00,0x00,0x00,0x03,0x00,0x00,0x04,0x00,0x00,0x18,0x00,0x00,0x20,0x00,0xF8,0x7F, // 78
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 79
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x10,0x01,0x00,0xE0, // 80
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x08,0x50,0x00,0x08,0x50,0x00,0x10,0x20,0x00,0x20,0x70,0x00,0xC0,0x4F, // 81
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x02,0x00,0x08,0x06,0x00,0x08,0x1A,0x00,0x10,0x21,0x00,0xE0,0x40, // 82
+  0x00,0x00,0x00,0x60,0x10,0x00,0x90,0x20,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x42,0x00,0x08,0x42,0x00,0x10,0x22,0x00,0x20,0x1C, // 83
+  0x08,0x00,0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08,0x00,0x00,0xF8,0x7F,0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08, // 84
+  0x00,0x00,0x00,0xF8,0x1F,0x00,0x00,0x20,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0xF8,0x1F, // 85
+  0x00,0x00,0x00,0x18,0x00,0x00,0xE0,0x00,0x00,0x00,0x07,0x00,0x00,0x18,0x00,0x00,0x60,0x00,0x00,0x18,0x00,0x00,0x07,0x00,0xE0,0x00,0x00,0x18, // 86
+  0x18,0x00,0x00,0xE0,0x01,0x00,0x00,0x1E,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x03,0x00,0x70,0x00,0x00,0x08,0x00,0x00,0x70,0x00,0x00,0x80,0x03,0x00,0x00,0x1C,0x00,0x00,0x60,0x00,0x00,0x1E,0x00,0xE0,0x01,0x00,0x18, // 87
+  0x00,0x40,0x00,0x08,0x20,0x00,0x10,0x10,0x00,0x60,0x0C,0x00,0x80,0x02,0x00,0x00,0x01,0x00,0x80,0x02,0x00,0x60,0x0C,0x00,0x10,0x10,0x00,0x08,0x20,0x00,0x00,0x40, // 88
+  0x08,0x00,0x00,0x30,0x00,0x00,0x40,0x00,0x00,0x80,0x01,0x00,0x00,0x7E,0x00,0x80,0x01,0x00,0x40,0x00,0x00,0x30,0x00,0x00,0x08, // 89
+  0x00,0x40,0x00,0x08,0x60,0x00,0x08,0x58,0x00,0x08,0x44,0x00,0x08,0x43,0x00,0x88,0x40,0x00,0x68,0x40,0x00,0x18,0x40,0x00,0x08,0x40, // 90
+  0x00,0x00,0x00,0xF8,0xFF,0x03,0x08,0x00,0x02,0x08,0x00,0x02, // 91
+  0x18,0x00,0x00,0xE0,0x01,0x00,0x00,0x1E,0x00,0x00,0x60, // 92
+  0x08,0x00,0x02,0x08,0x00,0x02,0xF8,0xFF,0x03, // 93
+  0x00,0x01,0x00,0xC0,0x00,0x00,0x30,0x00,0x00,0x08,0x00,0x00,0x30,0x00,0x00,0xC0,0x00,0x00,0x00,0x01, // 94
+  0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02, // 95
+  0x00,0x00,0x00,0x08,0x00,0x00,0x10, // 96
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x40,0x44,0x00,0x40,0x44,0x00,0x40,0x42,0x00,0x40,0x22,0x00,0x80,0x7F, // 97
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x80,0x20,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x80,0x20,0x00,0x00,0x1F, // 98
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x80,0x20, // 99
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x80,0x20,0x00,0xF8,0x7F, // 100
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x24,0x00,0x40,0x44,0x00,0x40,0x44,0x00,0x40,0x44,0x00,0x80,0x24,0x00,0x00,0x17, // 101
+  0x40,0x00,0x00,0xF0,0x7F,0x00,0x48,0x00,0x00,0x48, // 102
+  0x00,0x00,0x00,0x00,0x1F,0x01,0x80,0x20,0x02,0x40,0x40,0x02,0x40,0x40,0x02,0x40,0x40,0x02,0x80,0x20,0x01,0xC0,0xFF, // 103
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x80,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x80,0x7F, // 104
+  0x00,0x00,0x00,0xC8,0x7F, // 105
+  0x00,0x00,0x02,0xC8,0xFF,0x01, // 106
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x00,0x08,0x00,0x00,0x04,0x00,0x00,0x06,0x00,0x00,0x19,0x00,0x80,0x20,0x00,0x40,0x40, // 107
+  0x00,0x00,0x00,0xF8,0x7F, // 108
+  0x00,0x00,0x00,0xC0,0x7F,0x00,0x80,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x80,0x7F,0x00,0x80,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x80,0x7F, // 109
+  0x00,0x00,0x00,0xC0,0x7F,0x00,0x80,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x80,0x7F, // 110
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x80,0x20,0x00,0x00,0x1F, // 111
+  0x00,0x00,0x00,0xC0,0xFF,0x03,0x80,0x20,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x80,0x20,0x00,0x00,0x1F, // 112
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x80,0x20,0x00,0xC0,0xFF,0x03, // 113
+  0x00,0x00,0x00,0xC0,0x7F,0x00,0x80,0x00,0x00,0x40,0x00,0x00,0x40, // 114
+  0x00,0x00,0x00,0x80,0x23,0x00,0x40,0x44,0x00,0x40,0x44,0x00,0x40,0x44,0x00,0x40,0x44,0x00,0x80,0x38, // 115
+  0x40,0x00,0x00,0xF0,0x7F,0x00,0x40,0x40,0x00,0x40,0x40, // 116
+  0x00,0x00,0x00,0xC0,0x3F,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0xC0,0x7F, // 117
+  0xC0,0x00,0x00,0x00,0x03,0x00,0x00,0x1C,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x00,0x03,0x00,0xC0, // 118
+  0xC0,0x00,0x00,0x00,0x1F,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x00,0x03,0x00,0xC0,0x00,0x00,0x00,0x03,0x00,0x00,0x1C,0x00,0x00,0x60,0x00,0x00,0x1F,0x00,0xC0, // 119
+  0x40,0x40,0x00,0x80,0x20,0x00,0x00,0x1B,0x00,0x00,0x04,0x00,0x00,0x1B,0x00,0x80,0x20,0x00,0x40,0x40, // 120
+  0xC0,0x01,0x00,0x00,0x06,0x02,0x00,0x38,0x02,0x00,0xE0,0x01,0x00,0x38,0x00,0x00,0x07,0x00,0xC0, // 121
+  0x40,0x40,0x00,0x40,0x60,0x00,0x40,0x58,0x00,0x40,0x44,0x00,0x40,0x43,0x00,0xC0,0x40,0x00,0x40,0x40, // 122
+  0x00,0x04,0x00,0x00,0x04,0x00,0xF0,0xFB,0x01,0x08,0x00,0x02,0x08,0x00,0x02, // 123
+  0x00,0x00,0x00,0xF8,0xFF,0x03, // 124
+  0x08,0x00,0x02,0x08,0x00,0x02,0xF0,0xFB,0x01,0x00,0x04,0x00,0x00,0x04, // 125
+  0x00,0x02,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x00,0x01,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x01, // 126
+  0x00,0x00,0x00,0x60,0x10,0x00,0x90,0x20,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x0A,0x41,0x00,0x09,0x42,0x00,0x08,0x42,0x00,0x10,0x22,0x00,0x20,0x1C, // 140
+  0x00,0x40,0x00,0x08,0x60,0x00,0x08,0x58,0x00,0x08,0x44,0x00,0x08,0x43,0x00,0x8A,0x40,0x00,0x69,0x40,0x00,0x18,0x40,0x00,0x08,0x40, // 143
+  0x00,0x00,0x00,0x80,0x23,0x00,0x40,0x44,0x00,0x50,0x44,0x00,0x48,0x44,0x00,0x40,0x44,0x00,0x80,0x38, // 156
+  0x40,0x40,0x00,0x40,0x60,0x00,0x40,0x58,0x00,0x50,0x44,0x00,0x48,0x43,0x00,0xC0,0x40,0x00,0x40,0x40, // 159
+  0x00,0x00,0x00,0x00,0x00,0x00,0x40,0xFF,0x03, // 161
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x03,0x40,0xF0,0x00,0x40,0x4E,0x00,0xC0,0x41,0x00,0xB8,0x20,0x00,0x00,0x11, // 162
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x00,0x42,0x00,0x00,0x41,0x00,0x80,0x40,0x00,0x40,0x40,0x00,0x00,0x40,0x00,0x00,0x40, // 163
+  0x00,0x00,0x00,0x40,0x0B,0x00,0x80,0x04,0x00,0x40,0x08,0x00,0x40,0x08,0x00,0x80,0x04,0x00,0x40,0x0B, // 164
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x70,0x04,0x00,0x08,0x04,0x00,0x70,0x04,0x02,0x80,0x87,0x02,0x00,0x1C,0x03,0x00,0x60, // 165
+  0x00,0x00,0x00,0xF8,0xF1,0x03, // 166
+  0x00,0x86,0x00,0x70,0x09,0x01,0xC8,0x10,0x02,0x88,0x10,0x02,0x08,0x21,0x02,0x08,0x61,0x02,0x30,0xD2,0x01,0x00,0x0C, // 167
+  0x08,0x00,0x00,0x00,0x00,0x00,0x08, // 168
+  0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0xC8,0x47,0x00,0x28,0x48,0x00,0x28,0x48,0x00,0x28,0x48,0x00,0x28,0x48,0x00,0x48,0x44,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 169
+  0xD0,0x00,0x00,0x48,0x01,0x00,0x28,0x01,0x00,0x28,0x01,0x00,0xF0,0x01, // 170
+  0x00,0x00,0x00,0x00,0x04,0x00,0x00,0x1B,0x00,0x80,0x20,0x00,0x00,0x04,0x00,0x00,0x1B,0x00,0x80,0x20, // 171
+  0x00,0x00,0x00,0x80,0x00,0x00,0x80,0x00,0x00,0x80,0x00,0x00,0x80,0x00,0x00,0x80,0x00,0x00,0x80,0x00,0x00,0x80,0x0F, // 172
+  0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08,0x00,0x00,0x08, // 173
+  0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0xE8,0x4F,0x00,0x28,0x41,0x00,0x28,0x41,0x00,0x28,0x43,0x00,0x28,0x45,0x00,0xC8,0x48,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 174
+  0x00,0x40,0x00,0x08,0x60,0x00,0x08,0x58,0x00,0x08,0x44,0x00,0x0A,0x43,0x00,0x8A,0x40,0x00,0x68,0x40,0x00,0x18,0x40,0x00,0x08,0x40, // 175
+  0x00,0x00,0x00,0x30,0x00,0x00,0x48,0x00,0x00,0x48,0x00,0x00,0x30, // 176
+  0x00,0x00,0x00,0x00,0x41,0x00,0x00,0x41,0x00,0x00,0x41,0x00,0xE0,0x4F,0x00,0x00,0x41,0x00,0x00,0x41,0x00,0x00,0x41, // 177
+  0x10,0x01,0x00,0x88,0x01,0x00,0x48,0x01,0x00,0x48,0x01,0x00,0x30,0x01, // 178
+  0x00,0x06,0x00,0xF8,0x7F,0x00,0x80,0x01,0x00,0x40, // 179
+  0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x00,0x08, // 180
+  0x00,0x00,0x00,0xC0,0xFF,0x03,0x00,0x20,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0xC0,0x7F, // 181
+  0xF0,0x00,0x00,0xF8,0x00,0x00,0xF8,0x01,0x00,0xF8,0x01,0x00,0xF8,0xFF,0x03,0x08,0x00,0x00,0x08,0x00,0x00,0xF8,0xFF,0x03,0x08, // 182
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02, // 183
+  0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x80,0x02,0x00,0x00,0x03, // 184
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x40,0x44,0x00,0x40,0x44,0x02,0x40,0xC2,0x02,0x40,0x22,0x03,0x80,0x7F, // 185
+  0xF0,0x00,0x00,0x08,0x01,0x00,0x08,0x01,0x00,0x08,0x01,0x00,0xF0, // 186
+  0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x20,0x00,0x00,0x1B,0x00,0x00,0x04,0x00,0x80,0x20,0x00,0x00,0x1B,0x00,0x00,0x04, // 187
+  0x00,0x00,0x00,0x10,0x00,0x00,0x08,0x40,0x00,0xF8,0x21,0x00,0x00,0x10,0x00,0x00,0x0C,0x00,0x00,0x02,0x00,0x80,0x01,0x00,0x40,0x30,0x00,0x30,0x28,0x00,0x08,0x24,0x00,0x00,0x7E,0x00,0x00,0x20, // 188
+  0x00,0x00,0x00,0x10,0x00,0x00,0x08,0x40,0x00,0xF8,0x31,0x00,0x00,0x08,0x00,0x00,0x04,0x00,0x00,0x03,0x00,0x80,0x00,0x00,0x60,0x44,0x00,0x10,0x62,0x00,0x08,0x52,0x00,0x00,0x52,0x00,0x00,0x4C, // 189
+  0x90,0x00,0x00,0x08,0x01,0x00,0x08,0x41,0x00,0x28,0x21,0x00,0xD8,0x18,0x00,0x00,0x04,0x00,0x00,0x03,0x00,0x80,0x00,0x00,0x40,0x30,0x00,0x30,0x28,0x00,0x08,0x24,0x00,0x00,0x7E,0x00,0x00,0x20, // 190
+  0x40,0x40,0x00,0x40,0x60,0x00,0x40,0x58,0x00,0x50,0x44,0x00,0x50,0x43,0x00,0xC0,0x40,0x00,0x40,0x40, // 191
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x71,0x04,0x00,0x0A,0x04,0x00,0x70,0x04,0x00,0x80,0x07,0x00,0x00,0x1C,0x00,0x00,0x60, // 192
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x70,0x04,0x00,0x0A,0x04,0x00,0x71,0x04,0x00,0x80,0x07,0x00,0x00,0x1C,0x00,0x00,0x60, // 193
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x72,0x04,0x00,0x09,0x04,0x00,0x71,0x04,0x00,0x82,0x07,0x00,0x00,0x1C,0x00,0x00,0x60, // 194
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x72,0x04,0x00,0x09,0x04,0x00,0x72,0x04,0x00,0x81,0x07,0x00,0x00,0x1C,0x00,0x00,0x60, // 195
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x72,0x04,0x00,0x08,0x04,0x00,0x72,0x04,0x00,0x80,0x07,0x00,0x00,0x1C,0x00,0x00,0x60, // 196
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x1C,0x00,0x80,0x07,0x00,0x7E,0x04,0x00,0x0A,0x04,0x00,0x7E,0x04,0x00,0x80,0x07,0x00,0x00,0x1C,0x00,0x00,0x60, // 197
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x0A,0x40,0x00,0x09,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10, // 198
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x08,0x40,0x02,0x08,0xC0,0x02,0x08,0x40,0x03,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10, // 199
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x09,0x41,0x00,0x0A,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x40, // 200
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x0A,0x41,0x00,0x09,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x40, // 201
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x02,0x08,0xC1,0x02,0x08,0x41,0x03,0x08,0x41,0x00,0x08,0x40, // 202
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x0A,0x41,0x00,0x08,0x41,0x00,0x0A,0x41,0x00,0x08,0x41,0x00,0x08,0x41,0x00,0x08,0x40, // 203
+  0x01,0x00,0x00,0xFA,0x7F, // 204
+  0x00,0x00,0x00,0xFA,0x7F,0x00,0x01, // 205
+  0x02,0x00,0x00,0xF9,0x7F,0x00,0x01,0x00,0x00,0x02, // 206
+  0x02,0x00,0x00,0xF8,0x7F,0x00,0x02, // 207
+  0x00,0x02,0x00,0xF8,0x7F,0x00,0x08,0x42,0x00,0x08,0x42,0x00,0x08,0x42,0x00,0x08,0x42,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 208
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x10,0x00,0x00,0x60,0x00,0x00,0x80,0x00,0x00,0x00,0x03,0x00,0x02,0x04,0x00,0x01,0x18,0x00,0x00,0x20,0x00,0xF8,0x7F, // 209
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x09,0x40,0x00,0x0A,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 210
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x0A,0x40,0x00,0x09,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 211
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x0A,0x40,0x00,0x09,0x40,0x00,0x09,0x40,0x00,0x0A,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 212
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x0A,0x40,0x00,0x09,0x40,0x00,0x0A,0x40,0x00,0x09,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 213
+  0x00,0x00,0x00,0xC0,0x0F,0x00,0x20,0x10,0x00,0x10,0x20,0x00,0x08,0x40,0x00,0x0A,0x40,0x00,0x08,0x40,0x00,0x0A,0x40,0x00,0x10,0x20,0x00,0x20,0x10,0x00,0xC0,0x0F, // 214
+  0x00,0x00,0x00,0x40,0x10,0x00,0x80,0x08,0x00,0x00,0x05,0x00,0x00,0x07,0x00,0x00,0x05,0x00,0x80,0x08,0x00,0x40,0x10, // 215
+  0x00,0x00,0x00,0xC0,0x4F,0x00,0x20,0x30,0x00,0x10,0x30,0x00,0x08,0x4C,0x00,0x08,0x42,0x00,0x08,0x41,0x00,0xC8,0x40,0x00,0x30,0x20,0x00,0x30,0x10,0x00,0xC8,0x0F, // 216
+  0x00,0x00,0x00,0xF8,0x1F,0x00,0x00,0x20,0x00,0x00,0x40,0x00,0x01,0x40,0x00,0x02,0x40,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0xF8,0x1F, // 217
+  0x00,0x00,0x00,0xF8,0x1F,0x00,0x00,0x20,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x02,0x40,0x00,0x01,0x40,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0xF8,0x1F, // 218
+  0x00,0x00,0x00,0xF8,0x1F,0x00,0x00,0x20,0x00,0x00,0x40,0x00,0x02,0x40,0x00,0x01,0x40,0x00,0x01,0x40,0x00,0x02,0x40,0x00,0x00,0x20,0x00,0xF8,0x1F, // 219
+  0x00,0x00,0x00,0xF8,0x1F,0x00,0x00,0x20,0x00,0x00,0x40,0x00,0x02,0x40,0x00,0x00,0x40,0x00,0x02,0x40,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0xF8,0x1F, // 220
+  0x08,0x00,0x00,0x30,0x00,0x00,0x40,0x00,0x00,0x80,0x01,0x00,0x02,0x7E,0x00,0x81,0x01,0x00,0x40,0x00,0x00,0x30,0x00,0x00,0x08, // 221
+  0x00,0x00,0x00,0xF8,0x7F,0x00,0x20,0x10,0x00,0x20,0x10,0x00,0x20,0x10,0x00,0x20,0x10,0x00,0x20,0x10,0x00,0x20,0x10,0x00,0x40,0x08,0x00,0x80,0x07, // 222
+  0x00,0x00,0x00,0xE0,0x7F,0x00,0x10,0x00,0x00,0x08,0x20,0x00,0x88,0x43,0x00,0x70,0x42,0x00,0x00,0x44,0x00,0x00,0x38, // 223
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x40,0x44,0x00,0x48,0x44,0x00,0x50,0x42,0x00,0x40,0x22,0x00,0x80,0x7F, // 224
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x40,0x44,0x00,0x50,0x44,0x00,0x48,0x42,0x00,0x40,0x22,0x00,0x80,0x7F, // 225
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x50,0x44,0x00,0x48,0x44,0x00,0x48,0x42,0x00,0x50,0x22,0x00,0x80,0x7F, // 226
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x50,0x44,0x00,0x48,0x44,0x00,0x50,0x42,0x00,0x48,0x22,0x00,0x80,0x7F, // 227
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x50,0x44,0x00,0x40,0x44,0x00,0x50,0x42,0x00,0x40,0x22,0x00,0x80,0x7F, // 228
+  0x00,0x00,0x00,0x00,0x39,0x00,0x80,0x44,0x00,0x5C,0x44,0x00,0x54,0x44,0x00,0x5C,0x42,0x00,0x40,0x22,0x00,0x80,0x7F, // 229
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x40,0x40,0x00,0x50,0x40,0x00,0x48,0x40,0x00,0x80,0x20, // 230
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x40,0x40,0x02,0x40,0xC0,0x02,0x40,0x40,0x03,0x80,0x20, // 231
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x24,0x00,0x48,0x44,0x00,0x50,0x44,0x00,0x40,0x44,0x00,0x80,0x24,0x00,0x00,0x17, // 232
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x24,0x00,0x40,0x44,0x00,0x50,0x44,0x00,0x48,0x44,0x00,0x80,0x24,0x00,0x00,0x17, // 233
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x24,0x00,0x40,0x44,0x00,0x40,0x44,0x02,0x40,0xC4,0x02,0x80,0x24,0x03,0x00,0x17, // 234
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x24,0x00,0x50,0x44,0x00,0x40,0x44,0x00,0x50,0x44,0x00,0x80,0x24,0x00,0x00,0x17, // 235
+  0x08,0x00,0x00,0xD0,0x7F, // 236
+  0x00,0x00,0x00,0xD0,0x7F,0x00,0x08, // 237
+  0x10,0x00,0x00,0xC8,0x7F,0x00,0x08,0x00,0x00,0x10, // 238
+  0x10,0x00,0x00,0xC0,0x7F,0x00,0x10, // 239
+  0x00,0x00,0x00,0x00,0x1F,0x00,0xA0,0x20,0x00,0x68,0x40,0x00,0x58,0x40,0x00,0x70,0x40,0x00,0xE8,0x20,0x00,0x00,0x1F, // 240
+  0x00,0x00,0x00,0xC0,0x7F,0x00,0x80,0x00,0x00,0x40,0x00,0x00,0x50,0x00,0x00,0x48,0x00,0x00,0x80,0x7F, // 241
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x48,0x40,0x00,0x50,0x40,0x00,0x40,0x40,0x00,0x80,0x20,0x00,0x00,0x1F, // 242
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x40,0x40,0x00,0x50,0x40,0x00,0x48,0x40,0x00,0x80,0x20,0x00,0x00,0x1F, // 243
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x50,0x40,0x00,0x48,0x40,0x00,0x48,0x40,0x00,0x90,0x20,0x00,0x00,0x1F, // 244
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x50,0x40,0x00,0x48,0x40,0x00,0x50,0x40,0x00,0x88,0x20,0x00,0x00,0x1F, // 245
+  0x00,0x00,0x00,0x00,0x1F,0x00,0x80,0x20,0x00,0x50,0x40,0x00,0x40,0x40,0x00,0x50,0x40,0x00,0x80,0x20,0x00,0x00,0x1F, // 246
+  0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x80,0x0A,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02, // 247
+  0x00,0x00,0x00,0x00,0x5F,0x00,0x80,0x30,0x00,0x40,0x48,0x00,0x40,0x44,0x00,0x40,0x42,0x00,0x80,0x21,0x00,0x40,0x1F, // 248
+  0x00,0x00,0x00,0xC0,0x3F,0x00,0x00,0x40,0x00,0x08,0x40,0x00,0x10,0x40,0x00,0x00,0x20,0x00,0xC0,0x7F, // 249
+  0x00,0x00,0x00,0xC0,0x3F,0x00,0x00,0x40,0x00,0x00,0x40,0x00,0x10,0x40,0x00,0x08,0x20,0x00,0xC0,0x7F, // 250
+  0x00,0x00,0x00,0xC0,0x3F,0x00,0x10,0x40,0x00,0x08,0x40,0x00,0x08,0x40,0x00,0x10,0x20,0x00,0xC0,0x7F, // 251
+  0x00,0x00,0x00,0xD0,0x3F,0x00,0x00,0x40,0x00,0x10,0x40,0x00,0x00,0x40,0x00,0x00,0x20,0x00,0xC0,0x7F, // 252
+  0xC0,0x01,0x00,0x00,0x06,0x02,0x00,0x38,0x02,0x10,0xE0,0x01,0x08,0x38,0x00,0x00,0x07,0x00,0xC0, // 253
+  0x00,0x00,0x00,0xF8,0xFF,0x03,0x80,0x20,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x40,0x40,0x00,0x80,0x20,0x00,0x00,0x1F, // 254
+  0xC0,0x01,0x00,0x00,0x06,0x02,0x10,0x38,0x02,0x00,0xE0,0x01,0x10,0x38,0x00,0x00,0x07,0x00,0xC0 // 255
+};
+
+const uint8_t ArialMT_Win1250_Plain_24[] PROGMEM = {
+  0x18, // Width: 24
+  0x1C, // Height: 28
+  0x20, // First Char: 32
+  0xE0, // Numbers of Chars: 224
+
+  // Jump Table
+  0xFF, 0xFF, 0x00, 0x07, // 32:65535
+  0x00, 0x00, 0x13, 0x07, // 33:0
+  0x00, 0x13, 0x1A, 0x09, // 34:19
+  0x00, 0x2D, 0x33, 0x0D, // 35:45
+  0x00, 0x60, 0x2F, 0x0D, // 36:96
+  0x00, 0x8F, 0x4F, 0x15, // 37:143
+  0x00, 0xDE, 0x3B, 0x10, // 38:222
+  0x01, 0x19, 0x0A, 0x05, // 39:281
+  0x01, 0x23, 0x1C, 0x08, // 40:291
+  0x01, 0x3F, 0x1B, 0x08, // 41:319
+  0x01, 0x5A, 0x21, 0x09, // 42:346
+  0x01, 0x7B, 0x32, 0x0E, // 43:379
+  0x01, 0xAD, 0x10, 0x07, // 44:429
+  0x01, 0xBD, 0x1B, 0x08, // 45:445
+  0x01, 0xD8, 0x0F, 0x07, // 46:472
+  0x01, 0xE7, 0x19, 0x07, // 47:487
+  0x02, 0x00, 0x2F, 0x0D, // 48:512
+  0x02, 0x2F, 0x23, 0x0D, // 49:559
+  0x02, 0x52, 0x2F, 0x0D, // 50:594
+  0x02, 0x81, 0x2F, 0x0D, // 51:641
+  0x02, 0xB0, 0x2F, 0x0D, // 52:688
+  0x02, 0xDF, 0x2F, 0x0D, // 53:735
+  0x03, 0x0E, 0x2F, 0x0D, // 54:782
+  0x03, 0x3D, 0x2D, 0x0D, // 55:829
+  0x03, 0x6A, 0x2F, 0x0D, // 56:874
+  0x03, 0x99, 0x2F, 0x0D, // 57:921
+  0x03, 0xC8, 0x0F, 0x07, // 58:968
+  0x03, 0xD7, 0x10, 0x07, // 59:983
+  0x03, 0xE7, 0x2F, 0x0E, // 60:999
+  0x04, 0x16, 0x2F, 0x0E, // 61:1046
+  0x04, 0x45, 0x2E, 0x0E, // 62:1093
+  0x04, 0x73, 0x2E, 0x0D, // 63:1139
+  0x04, 0xA1, 0x5B, 0x18, // 64:1185
+  0x04, 0xFC, 0x3B, 0x10, // 65:1276
+  0x05, 0x37, 0x3B, 0x10, // 66:1335
+  0x05, 0x72, 0x3F, 0x11, // 67:1394
+  0x05, 0xB1, 0x3F, 0x11, // 68:1457
+  0x05, 0xF0, 0x3B, 0x10, // 69:1520
+  0x06, 0x2B, 0x35, 0x0F, // 70:1579
+  0x06, 0x60, 0x43, 0x13, // 71:1632
+  0x06, 0xA3, 0x3B, 0x11, // 72:1699
+  0x06, 0xDE, 0x0F, 0x07, // 73:1758
+  0x06, 0xED, 0x27, 0x0C, // 74:1773
+  0x07, 0x14, 0x3F, 0x10, // 75:1812
+  0x07, 0x53, 0x2F, 0x0D, // 76:1875
+  0x07, 0x82, 0x43, 0x14, // 77:1922
+  0x07, 0xC5, 0x3B, 0x11, // 78:1989
+  0x08, 0x00, 0x47, 0x13, // 79:2048
+  0x08, 0x47, 0x3A, 0x10, // 80:2119
+  0x08, 0x81, 0x47, 0x13, // 81:2177
+  0x08, 0xC8, 0x3F, 0x11, // 82:2248
+  0x09, 0x07, 0x3B, 0x10, // 83:2311
+  0x09, 0x42, 0x35, 0x0F, // 84:2370
+  0x09, 0x77, 0x3B, 0x11, // 85:2423
+  0x09, 0xB2, 0x39, 0x10, // 86:2482
+  0x09, 0xEB, 0x59, 0x17, // 87:2539
+  0x0A, 0x44, 0x3B, 0x10, // 88:2628
+  0x0A, 0x7F, 0x3D, 0x10, // 89:2687
+  0x0A, 0xBC, 0x37, 0x0F, // 90:2748
+  0x0A, 0xF3, 0x14, 0x07, // 91:2803
+  0x0B, 0x07, 0x1B, 0x07, // 92:2823
+  0x0B, 0x22, 0x18, 0x07, // 93:2850
+  0x0B, 0x3A, 0x2A, 0x0B, // 94:2874
+  0x0B, 0x64, 0x34, 0x0D, // 95:2916
+  0x0B, 0x98, 0x11, 0x08, // 96:2968
+  0x0B, 0xA9, 0x2F, 0x0D, // 97:2985
+  0x0B, 0xD8, 0x33, 0x0D, // 98:3032
+  0x0C, 0x0B, 0x2B, 0x0C, // 99:3083
+  0x0C, 0x36, 0x2F, 0x0D, // 100:3126
+  0x0C, 0x65, 0x2F, 0x0D, // 101:3173
+  0x0C, 0x94, 0x1A, 0x07, // 102:3220
+  0x0C, 0xAE, 0x2F, 0x0D, // 103:3246
+  0x0C, 0xDD, 0x2F, 0x0D, // 104:3293
+  0x0D, 0x0C, 0x0F, 0x05, // 105:3340
+  0x0D, 0x1B, 0x10, 0x05, // 106:3355
+  0x0D, 0x2B, 0x2F, 0x0C, // 107:3371
+  0x0D, 0x5A, 0x0F, 0x05, // 108:3418
+  0x0D, 0x69, 0x47, 0x14, // 109:3433
+  0x0D, 0xB0, 0x2F, 0x0D, // 110:3504
+  0x0D, 0xDF, 0x2F, 0x0D, // 111:3551
+  0x0E, 0x0E, 0x33, 0x0D, // 112:3598
+  0x0E, 0x41, 0x30, 0x0D, // 113:3649
+  0x0E, 0x71, 0x1E, 0x08, // 114:3697
+  0x0E, 0x8F, 0x2B, 0x0C, // 115:3727
+  0x0E, 0xBA, 0x1B, 0x07, // 116:3770
+  0x0E, 0xD5, 0x2F, 0x0D, // 117:3797
+  0x0F, 0x04, 0x2A, 0x0C, // 118:3844
+  0x0F, 0x2E, 0x42, 0x11, // 119:3886
+  0x0F, 0x70, 0x2B, 0x0C, // 120:3952
+  0x0F, 0x9B, 0x2A, 0x0C, // 121:3995
+  0x0F, 0xC5, 0x2B, 0x0C, // 122:4037
+  0x0F, 0xF0, 0x1C, 0x08, // 123:4080
+  0x10, 0x0C, 0x10, 0x06, // 124:4108
+  0x10, 0x1C, 0x1B, 0x08, // 125:4124
+  0x10, 0x37, 0x32, 0x0E, // 126:4151
+  0xFF, 0xFF, 0x00, 0x00, // 127:65535
+  0xFF, 0xFF, 0x00, 0x18, // 128:65535
+  0xFF, 0xFF, 0x00, 0x18, // 129:65535
+  0xFF, 0xFF, 0x00, 0x18, // 130:65535
+  0xFF, 0xFF, 0x00, 0x18, // 131:65535
+  0xFF, 0xFF, 0x00, 0x18, // 132:65535
+  0xFF, 0xFF, 0x00, 0x18, // 133:65535
+  0xFF, 0xFF, 0x00, 0x18, // 134:65535
+  0xFF, 0xFF, 0x00, 0x18, // 135:65535
+  0xFF, 0xFF, 0x00, 0x18, // 136:65535
+  0xFF, 0xFF, 0x00, 0x18, // 137:65535
+  0xFF, 0xFF, 0x00, 0x18, // 138:65535
+  0xFF, 0xFF, 0x00, 0x18, // 139:65535
+  0x10, 0x69, 0x3B, 0x10, // 140:4201
+  0xFF, 0xFF, 0x00, 0x18, // 141:65535
+  0xFF, 0xFF, 0x00, 0x18, // 142:65535
+  0x10, 0xA4, 0x37, 0x0F, // 143:4260
+  0xFF, 0xFF, 0x00, 0x18, // 144:65535
+  0xFF, 0xFF, 0x00, 0x18, // 145:65535
+  0xFF, 0xFF, 0x00, 0x18, // 146:65535
+  0xFF, 0xFF, 0x00, 0x18, // 147:65535
+  0xFF, 0xFF, 0x00, 0x18, // 148:65535
+  0xFF, 0xFF, 0x00, 0x18, // 149:65535
+  0xFF, 0xFF, 0x00, 0x18, // 150:65535
+  0xFF, 0xFF, 0x00, 0x18, // 151:65535
+  0xFF, 0xFF, 0x00, 0x18, // 152:65535
+  0xFF, 0xFF, 0x00, 0x18, // 153:65535
+  0xFF, 0xFF, 0x00, 0x18, // 154:65535
+  0xFF, 0xFF, 0x00, 0x18, // 155:65535
+  0x10, 0xDB, 0x2B, 0x0C, // 156:4315
+  0xFF, 0xFF, 0x00, 0x18, // 157:65535
+  0xFF, 0xFF, 0x00, 0x18, // 158:65535
+  0x11, 0x06, 0x2B, 0x0C, // 159:4358
+  0xFF, 0xFF, 0x00, 0x07, // 160:65535
+  0x11, 0x31, 0x14, 0x08, // 161:4401
+  0x11, 0x45, 0x2B, 0x0D, // 162:4421
+  0x11, 0x70, 0x2F, 0x0D, // 163:4464
+  0x11, 0x9F, 0x33, 0x0D, // 164:4511
+  0x11, 0xD2, 0x3B, 0x10, // 165:4562
+  0x12, 0x0D, 0x10, 0x06, // 166:4621
+  0x12, 0x1D, 0x2F, 0x0D, // 167:4637
+  0x12, 0x4C, 0x19, 0x08, // 168:4684
+  0x12, 0x65, 0x46, 0x12, // 169:4709
+  0x12, 0xAB, 0x1A, 0x09, // 170:4779
+  0x12, 0xC5, 0x27, 0x0D, // 171:4805
+  0x12, 0xEC, 0x2F, 0x0E, // 172:4844
+  0x13, 0x1B, 0x1B, 0x08, // 173:4891
+  0x13, 0x36, 0x46, 0x12, // 174:4918
+  0x13, 0x7C, 0x37, 0x0F, // 175:4988
+  0x13, 0xB3, 0x1E, 0x0A, // 176:5043
+  0x13, 0xD1, 0x33, 0x0D, // 177:5073
+  0x14, 0x04, 0x1A, 0x08, // 178:5124
+  0x14, 0x1E, 0x16, 0x06, // 179:5150
+  0x14, 0x34, 0x19, 0x08, // 180:5172
+  0x14, 0x4D, 0x2F, 0x0E, // 181:5197
+  0x14, 0x7C, 0x31, 0x0D, // 182:5244
+  0x14, 0xAD, 0x12, 0x08, // 183:5293
+  0x14, 0xBF, 0x18, 0x08, // 184:5311
+  0x14, 0xD7, 0x2F, 0x0D, // 185:5335
+  0x15, 0x06, 0x1E, 0x09, // 186:5382
+  0x15, 0x24, 0x2E, 0x0D, // 187:5412
+  0x15, 0x52, 0x4F, 0x14, // 188:5458
+  0x15, 0xA1, 0x4B, 0x14, // 189:5537
+  0x15, 0xEC, 0x4B, 0x14, // 190:5612
+  0x16, 0x37, 0x2B, 0x0C, // 191:5687
+  0x16, 0x62, 0x3B, 0x10, // 192:5730
+  0x16, 0x9D, 0x3B, 0x10, // 193:5789
+  0x16, 0xD8, 0x3B, 0x10, // 194:5848
+  0x17, 0x13, 0x3B, 0x10, // 195:5907
+  0x17, 0x4E, 0x3B, 0x10, // 196:5966
+  0x17, 0x89, 0x3B, 0x10, // 197:6025
+  0x17, 0xC4, 0x3F, 0x11, // 198:6084
+  0x18, 0x03, 0x3F, 0x11, // 199:6147
+  0x18, 0x42, 0x3B, 0x10, // 200:6210
+  0x18, 0x7D, 0x3B, 0x10, // 201:6269
+  0x18, 0xB8, 0x3B, 0x10, // 202:6328
+  0x18, 0xF3, 0x3B, 0x10, // 203:6387
+  0x19, 0x2E, 0x11, 0x07, // 204:6446
+  0x19, 0x3F, 0x11, 0x07, // 205:6463
+  0x19, 0x50, 0x15, 0x07, // 206:6480
+  0x19, 0x65, 0x15, 0x07, // 207:6501
+  0x19, 0x7A, 0x3F, 0x11, // 208:6522
+  0x19, 0xB9, 0x3B, 0x11, // 209:6585
+  0x19, 0xF4, 0x47, 0x13, // 210:6644
+  0x1A, 0x3B, 0x47, 0x13, // 211:6715
+  0x1A, 0x82, 0x47, 0x13, // 212:6786
+  0x1A, 0xC9, 0x47, 0x13, // 213:6857
+  0x1B, 0x10, 0x47, 0x13, // 214:6928
+  0x1B, 0x57, 0x2B, 0x0E, // 215:6999
+  0x1B, 0x82, 0x47, 0x13, // 216:7042
+  0x1B, 0xC9, 0x3B, 0x11, // 217:7113
+  0x1C, 0x04, 0x3B, 0x11, // 218:7172
+  0x1C, 0x3F, 0x3B, 0x11, // 219:7231
+  0x1C, 0x7A, 0x3B, 0x11, // 220:7290
+  0x1C, 0xB5, 0x3D, 0x10, // 221:7349
+  0x1C, 0xF2, 0x3A, 0x10, // 222:7410
+  0x1D, 0x2C, 0x37, 0x0F, // 223:7468
+  0x1D, 0x63, 0x2F, 0x0D, // 224:7523
+  0x1D, 0x92, 0x2F, 0x0D, // 225:7570
+  0x1D, 0xC1, 0x2F, 0x0D, // 226:7617
+  0x1D, 0xF0, 0x2F, 0x0D, // 227:7664
+  0x1E, 0x1F, 0x2F, 0x0D, // 228:7711
+  0x1E, 0x4E, 0x2F, 0x0D, // 229:7758
+  0x1E, 0x7D, 0x2B, 0x0C, // 230:7805
+  0x1E, 0xA8, 0x2B, 0x0C, // 231:7848
+  0x1E, 0xD3, 0x2F, 0x0D, // 232:7891
+  0x1F, 0x02, 0x2F, 0x0D, // 233:7938
+  0x1F, 0x31, 0x2F, 0x0D, // 234:7985
+  0x1F, 0x60, 0x2F, 0x0D, // 235:8032
+  0x1F, 0x8F, 0x11, 0x07, // 236:8079
+  0x1F, 0xA0, 0x11, 0x07, // 237:8096
+  0x1F, 0xB1, 0x15, 0x07, // 238:8113
+  0x1F, 0xC6, 0x15, 0x07, // 239:8134
+  0x1F, 0xDB, 0x2F, 0x0D, // 240:8155
+  0x20, 0x0A, 0x2F, 0x0D, // 241:8202
+  0x20, 0x39, 0x2F, 0x0D, // 242:8249
+  0x20, 0x68, 0x2F, 0x0D, // 243:8296
+  0x20, 0x97, 0x2F, 0x0D, // 244:8343
+  0x20, 0xC6, 0x2F, 0x0D, // 245:8390
+  0x20, 0xF5, 0x2F, 0x0D, // 246:8437
+  0x21, 0x24, 0x32, 0x0D, // 247:8484
+  0x21, 0x56, 0x33, 0x0F, // 248:8534
+  0x21, 0x89, 0x2F, 0x0D, // 249:8585
+  0x21, 0xB8, 0x2F, 0x0D, // 250:8632
+  0x21, 0xE7, 0x2F, 0x0D, // 251:8679
+  0x22, 0x16, 0x2F, 0x0D, // 252:8726
+  0x22, 0x45, 0x2A, 0x0C, // 253:8773
+  0x22, 0x6F, 0x2F, 0x0D, // 254:8815
+  0x22, 0x9E, 0x2A, 0x0C, // 255:8862
+
+  // Font Data
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x33,0x00,0xE0,0xFF,0x33, // 33
+  0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xE0,0x07,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xE0,0x07, // 34
+  0x00,0x0C,0x03,0x00,0x00,0x0C,0x33,0x00,0x00,0x0C,0x3F,0x00,0x00,0xFC,0x0F,0x00,0x80,0xFF,0x03,0x00,0xE0,0x0F,0x03,0x00,0x60,0x0C,0x33,0x00,0x00,0x0C,0x3F,0x00,0x00,0xFC,0x0F,0x00,0x80,0xFF,0x03,0x00,0xE0,0x0F,0x03,0x00,0x60,0x0C,0x03,0x00,0x00,0x0C,0x03, // 35
+  0x00,0x00,0x00,0x00,0x80,0x07,0x06,0x00,0xC0,0x0F,0x1E,0x00,0xC0,0x18,0x1C,0x00,0x60,0x18,0x38,0x00,0x60,0x30,0x30,0x00,0xF0,0xFF,0xFF,0x00,0x60,0x30,0x30,0x00,0x60,0x60,0x38,0x00,0xC0,0x60,0x18,0x00,0xC0,0xC1,0x1F,0x00,0x00,0x81,0x07, // 36
+  0x00,0x00,0x00,0x00,0x80,0x0F,0x00,0x00,0xC0,0x1F,0x00,0x00,0x60,0x30,0x00,0x00,0x20,0x20,0x00,0x00,0x20,0x20,0x20,0x00,0x60,0x30,0x38,0x00,0xC0,0x1F,0x1E,0x00,0x80,0x8F,0x0F,0x00,0x00,0xC0,0x03,0x00,0x00,0xF0,0x00,0x00,0x00,0x3C,0x00,0x00,0x00,0x8F,0x0F,0x00,0xC0,0xC3,0x1F,0x00,0xE0,0x60,0x30,0x00,0x20,0x20,0x20,0x00,0x00,0x20,0x20,0x00,0x00,0x60,0x30,0x00,0x00,0xC0,0x1F,0x00,0x00,0x80,0x0F, // 37
+  0x00,0x00,0x00,0x00,0x00,0x80,0x07,0x00,0x00,0xC0,0x0F,0x00,0x80,0xE3,0x1C,0x00,0xC0,0x77,0x38,0x00,0xE0,0x3C,0x30,0x00,0x60,0x38,0x30,0x00,0x60,0x78,0x30,0x00,0xE0,0xEC,0x38,0x00,0xC0,0x8F,0x1B,0x00,0x80,0x03,0x1F,0x00,0x00,0x00,0x0F,0x00,0x00,0xC0,0x1F,0x00,0x00,0xC0,0x38,0x00,0x00,0x00,0x10, // 38
+  0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xE0,0x07, // 39
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x0F,0x00,0x00,0xFE,0x7F,0x00,0x80,0x0F,0xF0,0x01,0xC0,0x01,0x80,0x03,0x60,0x00,0x00,0x06,0x20,0x00,0x00,0x04, // 40
+  0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x04,0x60,0x00,0x00,0x06,0xC0,0x01,0x80,0x03,0x80,0x0F,0xF0,0x01,0x00,0xFE,0x7F,0x00,0x00,0xF0,0x0F, // 41
+  0x00,0x00,0x00,0x00,0x80,0x00,0x00,0x00,0x80,0x04,0x00,0x00,0x80,0x0F,0x00,0x00,0xE0,0x03,0x00,0x00,0xE0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x80,0x04,0x00,0x00,0x80, // 42
+  0x00,0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0xFF,0x0F,0x00,0x00,0xFF,0x0F,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60, // 43
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x30,0x03,0x00,0x00,0xF0,0x01, // 44
+  0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01, // 45
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30, // 46
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0xE0,0x0F,0x00,0x00,0xFC,0x01,0x00,0x80,0x3F,0x00,0x00,0xE0,0x03,0x00,0x00,0x60, // 47
+  0x00,0x00,0x00,0x00,0x00,0xFE,0x03,0x00,0x80,0xFF,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xE0,0x00,0x38,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0xE0,0x00,0x38,0x00,0xC0,0x01,0x1C,0x00,0x80,0xFF,0x0F,0x00,0x00,0xFE,0x03, // 48
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x07,0x00,0x00,0x00,0x03,0x00,0x00,0x80,0x01,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 49
+  0x00,0x00,0x00,0x00,0x00,0x03,0x30,0x00,0xC0,0x03,0x38,0x00,0xC0,0x00,0x3C,0x00,0x60,0x00,0x36,0x00,0x60,0x00,0x33,0x00,0x60,0x80,0x31,0x00,0x60,0xC0,0x30,0x00,0x60,0x60,0x30,0x00,0xC0,0x30,0x30,0x00,0xC0,0x1F,0x30,0x00,0x00,0x0F,0x30, // 50
+  0x00,0x00,0x00,0x00,0x00,0x01,0x06,0x00,0xC0,0x01,0x0E,0x00,0xC0,0x00,0x1C,0x00,0x60,0x00,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0xC0,0x38,0x30,0x00,0xC0,0x6F,0x18,0x00,0x80,0xC7,0x0F,0x00,0x00,0x80,0x07, // 51
+  0x00,0x00,0x00,0x00,0x00,0x80,0x03,0x00,0x00,0xC0,0x03,0x00,0x00,0xF0,0x03,0x00,0x00,0x3C,0x03,0x00,0x00,0x0E,0x03,0x00,0x80,0x07,0x03,0x00,0xC0,0x01,0x03,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x00,0x03,0x00,0x00,0x00,0x03, // 52
+  0x00,0x00,0x00,0x00,0x00,0x30,0x06,0x00,0x80,0x3F,0x0E,0x00,0xE0,0x1F,0x18,0x00,0x60,0x08,0x30,0x00,0x60,0x0C,0x30,0x00,0x60,0x0C,0x30,0x00,0x60,0x0C,0x30,0x00,0x60,0x0C,0x30,0x00,0x60,0x18,0x1C,0x00,0x60,0xF0,0x0F,0x00,0x00,0xE0,0x03, // 53
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x03,0x00,0x80,0xFF,0x0F,0x00,0xC0,0x63,0x1C,0x00,0xC0,0x30,0x38,0x00,0x60,0x18,0x30,0x00,0x60,0x18,0x30,0x00,0x60,0x18,0x30,0x00,0x60,0x18,0x30,0x00,0xE0,0x30,0x18,0x00,0xC0,0xF1,0x0F,0x00,0x80,0xC1,0x07, // 54
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x3C,0x00,0x60,0x80,0x3F,0x00,0x60,0xE0,0x03,0x00,0x60,0x78,0x00,0x00,0x60,0x0E,0x00,0x00,0x60,0x03,0x00,0x00,0xE0,0x01,0x00,0x00,0x60, // 55
+  0x00,0x00,0x00,0x00,0x00,0x80,0x07,0x00,0x80,0xC7,0x1F,0x00,0xC0,0x6F,0x18,0x00,0xE0,0x38,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0xE0,0x38,0x30,0x00,0xC0,0x6F,0x18,0x00,0x80,0xC7,0x1F,0x00,0x00,0x80,0x07, // 56
+  0x00,0x00,0x00,0x00,0x00,0x1F,0x0C,0x00,0x80,0x7F,0x1C,0x00,0xC0,0x61,0x38,0x00,0x60,0xC0,0x30,0x00,0x60,0xC0,0x30,0x00,0x60,0xC0,0x30,0x00,0x60,0xC0,0x30,0x00,0x60,0x60,0x18,0x00,0xC0,0x31,0x1E,0x00,0x80,0xFF,0x0F,0x00,0x00,0xFE,0x01, // 57
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30, // 58
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x30,0x03,0x00,0x06,0xF0,0x01, // 59
+  0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0x50,0x00,0x00,0x00,0xD8,0x00,0x00,0x00,0xD8,0x00,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x04,0x01,0x00,0x00,0x06,0x03,0x00,0x00,0x06,0x03,0x00,0x00,0x03,0x06, // 60
+  0x00,0x00,0x00,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01, // 61
+  0x00,0x00,0x00,0x00,0x00,0x03,0x06,0x00,0x00,0x06,0x03,0x00,0x00,0x06,0x03,0x00,0x00,0x04,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0x8C,0x01,0x00,0x00,0xD8,0x00,0x00,0x00,0xD8,0x00,0x00,0x00,0x50,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0x20, // 62
+  0x00,0x00,0x00,0x00,0x00,0x03,0x00,0x00,0x80,0x03,0x00,0x00,0xC0,0x01,0x00,0x00,0xE0,0x00,0x00,0x00,0x60,0x80,0x33,0x00,0x60,0xC0,0x33,0x00,0x60,0xE0,0x00,0x00,0x60,0x30,0x00,0x00,0xC0,0x38,0x00,0x00,0xC0,0x1F,0x00,0x00,0x00,0x07, // 63
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x0F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x1E,0xF0,0x00,0x00,0x07,0xC0,0x01,0x80,0xC3,0x87,0x01,0xC0,0xF1,0x9F,0x03,0xC0,0x38,0x18,0x03,0xC0,0x0C,0x30,0x03,0x60,0x0E,0x30,0x06,0x60,0x06,0x30,0x06,0x60,0x06,0x18,0x06,0x60,0x06,0x0C,0x06,0x60,0x0C,0x1E,0x06,0x60,0xF8,0x3F,0x06,0xE0,0xFE,0x31,0x06,0xC0,0x0E,0x30,0x06,0xC0,0x01,0x18,0x03,0x80,0x03,0x1C,0x03,0x00,0x07,0x8F,0x01,0x00,0xFE,0x87,0x01,0x00,0xF8,0xC1,0x00,0x00,0x00,0x40, // 64
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x00,0xFE,0x01,0x00,0x80,0x8F,0x01,0x00,0xE0,0x83,0x01,0x00,0x60,0x80,0x01,0x00,0xE0,0x83,0x01,0x00,0x80,0x8F,0x01,0x00,0x00,0xFE,0x01,0x00,0x00,0xF0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 65
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0xC0,0x78,0x30,0x00,0xC0,0xFF,0x18,0x00,0x80,0xC7,0x1F,0x00,0x00,0x80,0x07, // 66
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x03,0x0F,0x00,0x00,0x02,0x03, // 67
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0xE0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x03,0x0E,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 68
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x00,0x30, // 69
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60, // 70
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0xE0,0x00,0x18,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x60,0x30,0x00,0x60,0x60,0x30,0x00,0xE0,0x60,0x38,0x00,0xC0,0x60,0x18,0x00,0xC0,0x61,0x18,0x00,0x80,0xE3,0x0F,0x00,0x00,0xE2,0x0F, // 71
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 72
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 73
+  0x00,0x00,0x00,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x1E,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x38,0x00,0xE0,0xFF,0x1F,0x00,0xE0,0xFF,0x0F, // 74
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0xE0,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x7C,0x00,0x00,0x00,0xFE,0x00,0x00,0x00,0xE7,0x01,0x00,0x80,0x83,0x07,0x00,0xC0,0x01,0x0F,0x00,0xE0,0x00,0x1E,0x00,0x60,0x00,0x38,0x00,0x20,0x00,0x30,0x00,0x00,0x00,0x20, // 75
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30, // 76
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0x01,0x00,0x00,0xC0,0x0F,0x00,0x00,0x00,0xFE,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0x00,0x3F,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x3F,0x00,0x00,0xE0,0x07,0x00,0x00,0xFE,0x00,0x00,0xC0,0x0F,0x00,0x00,0xE0,0x01,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 77
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0xC0,0x01,0x00,0x00,0x80,0x03,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x3C,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0xE0,0x01,0x00,0x00,0x80,0x03,0x00,0x00,0x00,0x0F,0x00,0x00,0x00,0x1C,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 78
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0xE0,0x00,0x38,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0xE0,0x00,0x38,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x07,0x0F,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 79
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x60,0x00,0x00,0x60,0x60,0x00,0x00,0x60,0x60,0x00,0x00,0x60,0x60,0x00,0x00,0x60,0x60,0x00,0x00,0x60,0x60,0x00,0x00,0x60,0x60,0x00,0x00,0x60,0x60,0x00,0x00,0xC0,0x30,0x00,0x00,0xC0,0x3F,0x00,0x00,0x00,0x0F, // 80
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x0C,0x00,0xC0,0x00,0x18,0x00,0xE0,0x00,0x18,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x36,0x00,0x60,0x00,0x36,0x00,0xE0,0x00,0x3C,0x00,0xC0,0x00,0x1C,0x00,0xC0,0x01,0x1C,0x00,0x80,0x07,0x3F,0x00,0x00,0xFF,0x77,0x00,0x00,0xFC,0x61, // 81
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x70,0x00,0x00,0x60,0xF0,0x00,0x00,0x60,0xF0,0x03,0x00,0x60,0xB0,0x07,0x00,0xE0,0x18,0x1F,0x00,0xC0,0x1F,0x3C,0x00,0x80,0x0F,0x30,0x00,0x00,0x00,0x20, // 82
+  0x00,0x00,0x00,0x00,0x00,0x00,0x03,0x00,0x00,0x07,0x0F,0x00,0xC0,0x1F,0x1C,0x00,0xC0,0x18,0x18,0x00,0x60,0x38,0x38,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x70,0x30,0x00,0xC0,0x60,0x18,0x00,0xC0,0xE1,0x18,0x00,0x80,0xC3,0x0F,0x00,0x00,0x83,0x07, // 83
+  0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60, // 84
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x03,0x00,0xE0,0xFF,0x0F,0x00,0x00,0x00,0x1C,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x1C,0x00,0xE0,0xFF,0x0F,0x00,0xE0,0xFF,0x03, // 85
+  0x20,0x00,0x00,0x00,0xE0,0x01,0x00,0x00,0xC0,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0xF8,0x01,0x00,0x00,0xC0,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x3E,0x00,0x00,0xC0,0x0F,0x00,0x00,0xF8,0x01,0x00,0x00,0x3E,0x00,0x00,0xC0,0x0F,0x00,0x00,0xE0,0x01,0x00,0x00,0x20, // 86
+  0x60,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0x80,0xFF,0x00,0x00,0x00,0xF8,0x0F,0x00,0x00,0x80,0x3F,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x3F,0x00,0x00,0xE0,0x0F,0x00,0x00,0xFC,0x01,0x00,0x80,0x1F,0x00,0x00,0xE0,0x03,0x00,0x00,0x60,0x00,0x00,0x00,0xE0,0x03,0x00,0x00,0x80,0x1F,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xE0,0x0F,0x00,0x00,0x00,0x3F,0x00,0x00,0x00,0x30,0x00,0x00,0x80,0x3F,0x00,0x00,0xF8,0x0F,0x00,0x80,0xFF,0x00,0x00,0xE0,0x07,0x00,0x00,0x60, // 87
+  0x00,0x00,0x20,0x00,0x20,0x00,0x30,0x00,0x60,0x00,0x3C,0x00,0xE0,0x01,0x1E,0x00,0xC0,0x83,0x07,0x00,0x00,0xCF,0x03,0x00,0x00,0xFE,0x01,0x00,0x00,0x38,0x00,0x00,0x00,0xFE,0x01,0x00,0x00,0xCF,0x03,0x00,0xC0,0x03,0x07,0x00,0xE0,0x01,0x1E,0x00,0x60,0x00,0x3C,0x00,0x20,0x00,0x30,0x00,0x00,0x00,0x20, // 88
+  0x20,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0xC0,0x01,0x00,0x00,0x80,0x03,0x00,0x00,0x00,0x07,0x00,0x00,0x00,0x1E,0x00,0x00,0x00,0x3C,0x00,0x00,0x00,0xF0,0x3F,0x00,0x00,0xF0,0x3F,0x00,0x00,0x3C,0x00,0x00,0x00,0x1E,0x00,0x00,0x00,0x07,0x00,0x00,0xC0,0x03,0x00,0x00,0xE0,0x01,0x00,0x00,0x60,0x00,0x00,0x00,0x20, // 89
+  0x00,0x00,0x30,0x00,0x60,0x00,0x38,0x00,0x60,0x00,0x3C,0x00,0x60,0x00,0x37,0x00,0x60,0x80,0x33,0x00,0x60,0xC0,0x31,0x00,0x60,0xE0,0x30,0x00,0x60,0x38,0x30,0x00,0x60,0x1C,0x30,0x00,0x60,0x0E,0x30,0x00,0x60,0x07,0x30,0x00,0xE0,0x01,0x30,0x00,0xE0,0x00,0x30,0x00,0x60,0x00,0x30, // 90
+  0x00,0x00,0x00,0x00,0xE0,0xFF,0xFF,0x07,0xE0,0xFF,0xFF,0x07,0x60,0x00,0x00,0x06,0x60,0x00,0x00,0x06, // 91
+  0x60,0x00,0x00,0x00,0xE0,0x03,0x00,0x00,0x80,0x3F,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xE0,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 92
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x06,0x60,0x00,0x00,0x06,0xE0,0xFF,0xFF,0x07,0xE0,0xFF,0xFF,0x07, // 93
+  0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x1F,0x00,0x00,0xC0,0x07,0x00,0x00,0xE0,0x00,0x00,0x00,0xE0,0x00,0x00,0x00,0xC0,0x07,0x00,0x00,0x00,0x1F,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x20, // 94
+  0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06, // 95
+  0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0xE0,0x00,0x00,0x00,0x80, // 96
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0x00,0x1C,0x1F,0x00,0x00,0x8C,0x39,0x00,0x00,0x86,0x31,0x00,0x00,0x86,0x31,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x18,0x00,0x00,0xCE,0x0C,0x00,0x00,0xFC,0x1F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x00,0x20, // 97
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x18,0x0C,0x00,0x00,0x0C,0x18,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x0E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xE0,0x03, // 98
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x38,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x0E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0x18,0x0C, // 99
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x03,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x38,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x0C,0x18,0x00,0x00,0x18,0x0C,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 100
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0xDC,0x1C,0x00,0x00,0xCE,0x38,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x30,0x00,0x00,0xCE,0x38,0x00,0x00,0xDC,0x18,0x00,0x00,0xF8,0x0C,0x00,0x00,0xF0,0x04, // 101
+  0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0xC0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x06,0x00,0x00,0x60,0x06,0x00,0x00,0x60,0x06, // 102
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x83,0x01,0x00,0xF8,0x8F,0x03,0x00,0x1C,0x1C,0x07,0x00,0x0E,0x38,0x06,0x00,0x06,0x30,0x06,0x00,0x06,0x30,0x06,0x00,0x06,0x30,0x06,0x00,0x0C,0x18,0x07,0x00,0x18,0x8C,0x03,0x00,0xFE,0xFF,0x01,0x00,0xFE,0xFF, // 103
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x18,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0xFC,0x3F,0x00,0x00,0xF8,0x3F, // 104
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x60,0xFE,0x3F,0x00,0x60,0xFE,0x3F, // 105
+  0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x60,0xFE,0xFF,0x07,0x60,0xFE,0xFF,0x03, // 106
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0xC0,0x00,0x00,0x00,0xE0,0x00,0x00,0x00,0xF0,0x01,0x00,0x00,0x98,0x07,0x00,0x00,0x0C,0x0E,0x00,0x00,0x06,0x3C,0x00,0x00,0x02,0x30,0x00,0x00,0x00,0x20, // 107
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 108
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F,0x00,0x00,0x0C,0x00,0x00,0x00,0x04,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0xFC,0x3F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x0C,0x00,0x00,0x00,0x04,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0xFC,0x3F,0x00,0x00,0xF8,0x3F, // 109
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F,0x00,0x00,0x18,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0xFC,0x3F,0x00,0x00,0xF8,0x3F, // 110
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x38,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x0E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xF0,0x07, // 111
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0xFF,0x07,0x00,0xFE,0xFF,0x07,0x00,0x18,0x0C,0x00,0x00,0x0C,0x18,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x0E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xE0,0x03, // 112
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x03,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x38,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x0C,0x18,0x00,0x00,0x18,0x0C,0x00,0x00,0xFE,0xFF,0x07,0x00,0xFE,0xFF,0x07, // 113
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F,0x00,0x00,0x0C,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0x00,0x06, // 114
+  0x00,0x00,0x00,0x00,0x00,0x38,0x0C,0x00,0x00,0x7C,0x1C,0x00,0x00,0xEE,0x38,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x31,0x00,0x00,0xC6,0x31,0x00,0x00,0x8E,0x39,0x00,0x00,0x9C,0x1F,0x00,0x00,0x18,0x0F, // 115
+  0x00,0x06,0x00,0x00,0x00,0x06,0x00,0x00,0xC0,0xFF,0x1F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30, // 116
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x0F,0x00,0x00,0xFE,0x1F,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x18,0x00,0x00,0x00,0x0C,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F, // 117
+  0x00,0x06,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0xF8,0x00,0x00,0x00,0xC0,0x07,0x00,0x00,0x00,0x1F,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x1F,0x00,0x00,0xC0,0x07,0x00,0x00,0xF8,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x06, // 118
+  0x00,0x0E,0x00,0x00,0x00,0x7E,0x00,0x00,0x00,0xF0,0x03,0x00,0x00,0x80,0x1F,0x00,0x00,0x00,0x38,0x00,0x00,0x80,0x1F,0x00,0x00,0xE0,0x03,0x00,0x00,0x7C,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x7C,0x00,0x00,0x00,0xE0,0x03,0x00,0x00,0x80,0x1F,0x00,0x00,0x00,0x38,0x00,0x00,0x80,0x1F,0x00,0x00,0xF0,0x03,0x00,0x00,0x7E,0x00,0x00,0x00,0x0E, // 119
+  0x00,0x02,0x20,0x00,0x00,0x06,0x30,0x00,0x00,0x1E,0x3C,0x00,0x00,0x38,0x0E,0x00,0x00,0xF0,0x07,0x00,0x00,0xC0,0x01,0x00,0x00,0xE0,0x07,0x00,0x00,0x38,0x0E,0x00,0x00,0x1C,0x3C,0x00,0x00,0x0E,0x30,0x00,0x00,0x02,0x20, // 120
+  0x00,0x00,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x7E,0x00,0x06,0x00,0xF0,0x01,0x06,0x00,0x80,0x0F,0x07,0x00,0x00,0xFE,0x03,0x00,0x00,0xFC,0x00,0x00,0xC0,0x1F,0x00,0x00,0xF8,0x03,0x00,0x00,0x3E,0x00,0x00,0x00,0x06, // 121
+  0x00,0x00,0x00,0x00,0x00,0x00,0x30,0x00,0x00,0x06,0x3C,0x00,0x00,0x06,0x3E,0x00,0x00,0x06,0x37,0x00,0x00,0xC6,0x33,0x00,0x00,0xE6,0x30,0x00,0x00,0x76,0x30,0x00,0x00,0x3E,0x30,0x00,0x00,0x1E,0x30,0x00,0x00,0x06,0x30, // 122
+  0x00,0x00,0x00,0x00,0x00,0x80,0x01,0x00,0x00,0xC0,0x03,0x00,0xC0,0x7F,0xFE,0x03,0xE0,0x3F,0xFC,0x07,0x60,0x00,0x00,0x06,0x60,0x00,0x00,0x06, // 123
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0xFF,0x0F,0xE0,0xFF,0xFF,0x0F, // 124
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x06,0x60,0x00,0x00,0x06,0xE0,0x3F,0xFC,0x07,0xC0,0x7F,0xFF,0x03,0x00,0xC0,0x03,0x00,0x00,0x80,0x01, // 125
+  0x00,0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0xC0,0x00,0x00,0x00,0xC0,0x00,0x00,0x00,0xC0,0x00,0x00,0x00,0xE0,0x00,0x00,0x00,0x60, // 126
+  0x00,0x00,0x00,0x00,0x00,0x00,0x03,0x00,0x00,0x07,0x0F,0x00,0xC0,0x1F,0x1C,0x00,0xC0,0x18,0x18,0x00,0x60,0x38,0x38,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x68,0x30,0x30,0x00,0x6E,0x30,0x30,0x00,0x66,0x70,0x30,0x00,0xC2,0x60,0x18,0x00,0xC0,0xE1,0x18,0x00,0x80,0xC3,0x0F,0x00,0x00,0x83,0x07, // 140
+  0x00,0x00,0x30,0x00,0x60,0x00,0x38,0x00,0x60,0x00,0x3C,0x00,0x60,0x00,0x37,0x00,0x60,0x80,0x33,0x00,0x60,0xC0,0x31,0x00,0x60,0xE0,0x30,0x00,0x60,0x38,0x30,0x00,0x68,0x1C,0x30,0x00,0x6E,0x0E,0x30,0x00,0x66,0x07,0x30,0x00,0xE2,0x01,0x30,0x00,0xE0,0x00,0x30,0x00,0x60,0x00,0x30, // 143
+  0x00,0x00,0x00,0x00,0x00,0x38,0x0C,0x00,0x00,0x7C,0x1C,0x00,0x00,0xEE,0x38,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x30,0x00,0x80,0xC6,0x31,0x00,0xE0,0xC6,0x31,0x00,0x60,0x8E,0x39,0x00,0x20,0x9C,0x1F,0x00,0x00,0x18,0x0F, // 156
+  0x00,0x00,0x00,0x00,0x00,0x00,0x30,0x00,0x00,0x06,0x3C,0x00,0x00,0x06,0x3E,0x00,0x00,0x06,0x37,0x00,0x00,0xC6,0x33,0x00,0x80,0xE6,0x30,0x00,0xE0,0x76,0x30,0x00,0x60,0x3E,0x30,0x00,0x20,0x1E,0x30,0x00,0x00,0x06,0x30, // 159
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE6,0xFF,0x07,0x00,0xE6,0xFF,0x07, // 161
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x9C,0x07,0x00,0x0E,0x78,0x00,0x00,0x06,0x3F,0x00,0x00,0xF6,0x30,0x00,0x00,0x0E,0x30,0x00,0xE0,0x0D,0x1C,0x00,0x00,0x1C,0x0E,0x00,0x00,0x10,0x06, // 162
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x30,0x30,0x00,0x00,0x18,0x30,0x00,0x00,0x0C,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x03,0x30,0x00,0x00,0x01,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30, // 163
+  0x00,0x00,0x00,0x00,0x00,0x02,0x04,0x00,0x00,0xF7,0x0E,0x00,0x00,0xFE,0x07,0x00,0x00,0x0C,0x03,0x00,0x00,0x06,0x06,0x00,0x00,0x06,0x06,0x00,0x00,0x06,0x06,0x00,0x00,0x06,0x06,0x00,0x00,0x0C,0x03,0x00,0x00,0xFE,0x07,0x00,0x00,0xF7,0x0E,0x00,0x00,0x02,0x04, // 164
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x00,0xFE,0x01,0x00,0x80,0x8F,0x01,0x00,0xE0,0x83,0x01,0x00,0x60,0x80,0x01,0x00,0xE0,0x83,0x01,0x02,0x80,0x8F,0x01,0x02,0x00,0xFE,0xC1,0x02,0x00,0xF0,0x83,0x03,0x00,0x80,0x0F,0x01,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 165
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0x7F,0xF8,0x0F,0xE0,0x7F,0xF8,0x0F, // 166
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x00,0x00,0x80,0xF3,0xC1,0x00,0xC0,0x1F,0xC3,0x03,0xE0,0x0C,0x07,0x03,0x60,0x1C,0x06,0x06,0x60,0x18,0x0C,0x06,0x60,0x30,0x1C,0x06,0xE0,0x70,0x38,0x07,0xC0,0xE1,0xF4,0x03,0x80,0xC1,0xE7,0x01,0x00,0x80,0x03, // 167
+  0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60, // 168
+  0x00,0xF8,0x00,0x00,0x00,0xFE,0x03,0x00,0x00,0x07,0x07,0x00,0x80,0x01,0x0C,0x00,0xC0,0x79,0x1C,0x00,0xC0,0xFE,0x19,0x00,0x60,0x86,0x31,0x00,0x60,0x03,0x33,0x00,0x60,0x03,0x33,0x00,0x60,0x03,0x33,0x00,0x60,0x03,0x33,0x00,0x60,0x87,0x33,0x00,0xC0,0x86,0x19,0x00,0xC0,0x85,0x1C,0x00,0x80,0x01,0x0C,0x00,0x00,0x07,0x07,0x00,0x00,0xFE,0x03,0x00,0x00,0xF8, // 169
+  0x00,0x00,0x00,0x00,0xC0,0x1C,0x00,0x00,0xE0,0x3E,0x00,0x00,0x60,0x32,0x00,0x00,0x60,0x32,0x00,0x00,0xE0,0x3F,0x00,0x00,0xC0,0x3F, // 170
+  0x00,0x00,0x00,0x00,0x00,0x80,0x00,0x00,0x00,0xE0,0x03,0x00,0x00,0x78,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x84,0x10,0x00,0x00,0xE0,0x03,0x00,0x00,0x78,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x04,0x10, // 171
+  0x00,0x00,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFC,0x01, // 172
+  0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01,0x00,0x00,0x80,0x01, // 173
+  0x00,0xF8,0x00,0x00,0x00,0xFE,0x03,0x00,0x00,0x07,0x07,0x00,0x80,0x01,0x0C,0x00,0xC0,0x01,0x1C,0x00,0xC0,0xFE,0x1B,0x00,0x60,0xFE,0x33,0x00,0x60,0x66,0x30,0x00,0x60,0x66,0x30,0x00,0x60,0xE6,0x30,0x00,0x60,0xFE,0x31,0x00,0x60,0x3C,0x33,0x00,0xC0,0x00,0x1A,0x00,0xC0,0x01,0x1C,0x00,0x80,0x01,0x0C,0x00,0x00,0x07,0x07,0x00,0x00,0xFE,0x03,0x00,0x00,0xF8, // 174
+  0x00,0x00,0x30,0x00,0x60,0x00,0x38,0x00,0x60,0x00,0x3C,0x00,0x60,0x00,0x37,0x00,0x60,0x80,0x33,0x00,0x60,0xC0,0x31,0x00,0x6C,0xE0,0x30,0x00,0x6C,0x38,0x30,0x00,0x60,0x1C,0x30,0x00,0x60,0x0E,0x30,0x00,0x60,0x07,0x30,0x00,0xE0,0x01,0x30,0x00,0xE0,0x00,0x30,0x00,0x60,0x00,0x30, // 175
+  0x00,0x00,0x00,0x00,0x80,0x03,0x00,0x00,0x40,0x04,0x00,0x00,0x20,0x08,0x00,0x00,0x20,0x08,0x00,0x00,0x20,0x08,0x00,0x00,0x40,0x04,0x00,0x00,0x80,0x03, // 176
+  0x00,0x00,0x00,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0xFF,0x3F,0x00,0x00,0xFF,0x3F,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30,0x00,0x00,0x60,0x30, // 177
+  0x40,0x20,0x00,0x00,0x60,0x30,0x00,0x00,0x20,0x38,0x00,0x00,0x20,0x2C,0x00,0x00,0x20,0x26,0x00,0x00,0xE0,0x23,0x00,0x00,0xC0,0x21, // 178
+  0x00,0x80,0x01,0x00,0x00,0xC0,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x18,0x00,0x00,0x00,0x0C, // 179
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x00,0x00,0x00,0xE0,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x20, // 180
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0xFF,0x07,0x00,0xFE,0xFF,0x07,0x00,0x00,0x1C,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x1C,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F, // 181
+  0x00,0x0F,0x00,0x00,0xC0,0x3F,0x00,0x00,0xC0,0x3F,0x00,0x00,0xE0,0x7F,0x00,0x00,0xE0,0x7F,0x00,0x00,0xE0,0xFF,0xFF,0x07,0xE0,0xFF,0xFF,0x07,0x60,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0xE0,0xFF,0xFF,0x07,0xE0,0xFF,0xFF,0x07,0x60,0x00,0x00,0x00,0x60, // 182
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0x60, // 183
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x02,0x00,0x00,0xC0,0x02,0x00,0x00,0x80,0x03,0x00,0x00,0x00,0x01, // 184
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0x00,0x1C,0x1F,0x00,0x00,0x8C,0x39,0x00,0x00,0x86,0x31,0x00,0x00,0x86,0x31,0x00,0x00,0xC6,0x30,0x02,0x00,0xC6,0x18,0x02,0x00,0xCE,0xCC,0x02,0x00,0xFC,0x9F,0x03,0x00,0xF8,0x3F,0x01,0x00,0x00,0x20, // 185
+  0x00,0x00,0x00,0x00,0x80,0x0F,0x00,0x00,0xC0,0x1F,0x00,0x00,0xE0,0x38,0x00,0x00,0x60,0x30,0x00,0x00,0xE0,0x38,0x00,0x00,0xC0,0x1F,0x00,0x00,0x80,0x0F, // 186
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x10,0x00,0x00,0x1C,0x1C,0x00,0x00,0x78,0x0F,0x00,0x00,0xE0,0x03,0x00,0x00,0x84,0x10,0x00,0x00,0x1C,0x1C,0x00,0x00,0x78,0x0F,0x00,0x00,0xE0,0x03,0x00,0x00,0x80, // 187
+  0x00,0x00,0x00,0x00,0x80,0x01,0x00,0x00,0xC0,0x00,0x00,0x00,0xC0,0x00,0x20,0x00,0xE0,0x3F,0x38,0x00,0xE0,0x3F,0x1C,0x00,0x00,0x00,0x0E,0x00,0x00,0x80,0x03,0x00,0x00,0xC0,0x01,0x00,0x00,0xE0,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x1C,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x07,0x0C,0x00,0xC0,0x01,0x0E,0x00,0xE0,0x80,0x0B,0x00,0x60,0xC0,0x08,0x00,0x00,0xE0,0x3F,0x00,0x00,0xE0,0x3F,0x00,0x00,0x00,0x08, // 188
+  0x00,0x00,0x00,0x00,0x80,0x01,0x00,0x00,0xC0,0x00,0x00,0x00,0xC0,0x00,0x20,0x00,0xE0,0x3F,0x30,0x00,0xE0,0x3F,0x1C,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x07,0x00,0x00,0xC0,0x01,0x00,0x00,0xE0,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0x1C,0x00,0x00,0x00,0x4E,0x20,0x00,0x00,0x67,0x30,0x00,0xC0,0x21,0x38,0x00,0xE0,0x20,0x2C,0x00,0x60,0x20,0x26,0x00,0x00,0xE0,0x27,0x00,0x00,0xC0,0x21, // 189
+  0x40,0x10,0x00,0x00,0x60,0x30,0x00,0x00,0x20,0x20,0x00,0x00,0x20,0x22,0x20,0x00,0x20,0x22,0x30,0x00,0xE0,0x3D,0x38,0x00,0xC0,0x1D,0x0E,0x00,0x00,0x00,0x07,0x00,0x00,0x80,0x03,0x00,0x00,0xE0,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x0E,0x0C,0x00,0x00,0x07,0x0E,0x00,0x80,0x83,0x0B,0x00,0xE0,0xC0,0x08,0x00,0x60,0xE0,0x3F,0x00,0x20,0xE0,0x3F,0x00,0x00,0x00,0x08, // 190
+  0x00,0x00,0x00,0x00,0x00,0x00,0x30,0x00,0x00,0x06,0x3C,0x00,0x00,0x06,0x3E,0x00,0x00,0x06,0x37,0x00,0x00,0xC6,0x33,0x00,0xC0,0xE6,0x30,0x00,0xC0,0x76,0x30,0x00,0x00,0x3E,0x30,0x00,0x00,0x1E,0x30,0x00,0x00,0x06,0x30, // 191
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x00,0xFE,0x01,0x00,0x82,0x8F,0x01,0x00,0xE6,0x83,0x01,0x00,0x6E,0x80,0x01,0x00,0xE8,0x83,0x01,0x00,0x80,0x8F,0x01,0x00,0x00,0xFE,0x01,0x00,0x00,0xF0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 192
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x00,0xFE,0x01,0x00,0x80,0x8F,0x01,0x00,0xE8,0x83,0x01,0x00,0x6E,0x80,0x01,0x00,0xE6,0x83,0x01,0x00,0x82,0x8F,0x01,0x00,0x00,0xFE,0x01,0x00,0x00,0xF0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 193
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x00,0xFE,0x01,0x00,0x88,0x8F,0x01,0x00,0xEC,0x83,0x01,0x00,0x66,0x80,0x01,0x00,0xE6,0x83,0x01,0x00,0x8C,0x8F,0x01,0x00,0x08,0xFE,0x01,0x00,0x00,0xF0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 194
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x0C,0xFE,0x01,0x00,0x8E,0x8F,0x01,0x00,0xE6,0x83,0x01,0x00,0x66,0x80,0x01,0x00,0xEC,0x83,0x01,0x00,0x8C,0x8F,0x01,0x00,0x0E,0xFE,0x01,0x00,0x06,0xF0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 195
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x00,0xFE,0x01,0x00,0x8C,0x8F,0x01,0x00,0xEC,0x83,0x01,0x00,0x60,0x80,0x01,0x00,0xE0,0x83,0x01,0x00,0x8C,0x8F,0x01,0x00,0x0C,0xFE,0x01,0x00,0x00,0xF0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 196
+  0x00,0x00,0x30,0x00,0x00,0x00,0x3E,0x00,0x00,0x80,0x0F,0x00,0x00,0xF0,0x03,0x00,0x00,0xFE,0x01,0x00,0x9C,0x8F,0x01,0x00,0xE2,0x83,0x01,0x00,0x62,0x80,0x01,0x00,0xE2,0x83,0x01,0x00,0x9C,0x8F,0x01,0x00,0x00,0xFE,0x01,0x00,0x00,0xF0,0x03,0x00,0x00,0x80,0x0F,0x00,0x00,0x00,0x3E,0x00,0x00,0x00,0x30, // 197
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x68,0x00,0x30,0x00,0x6E,0x00,0x30,0x00,0x66,0x00,0x30,0x00,0x62,0x00,0x30,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x03,0x0F,0x00,0x00,0x02,0x03, // 198
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0x60,0x00,0x30,0x02,0x60,0x00,0x30,0x02,0x60,0x00,0xF0,0x02,0x60,0x00,0xB0,0x03,0x60,0x00,0x30,0x01,0x60,0x00,0x30,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x03,0x0F,0x00,0x00,0x02,0x03, // 199
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x62,0x30,0x30,0x00,0x66,0x30,0x30,0x00,0x6E,0x30,0x30,0x00,0x68,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x00,0x30, // 200
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x68,0x30,0x30,0x00,0x6E,0x30,0x30,0x00,0x66,0x30,0x30,0x00,0x62,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x00,0x30, // 201
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x02,0x60,0x30,0x30,0x02,0x60,0x30,0xF0,0x02,0x60,0x30,0xB0,0x03,0x60,0x30,0x30,0x01,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x00,0x30, // 202
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x6C,0x30,0x30,0x00,0x6C,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x6C,0x30,0x30,0x00,0x6C,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x00,0x30, // 203
+  0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0xE6,0xFF,0x3F,0x00,0xEE,0xFF,0x3F,0x00,0x08, // 204
+  0x00,0x00,0x00,0x00,0x08,0x00,0x00,0x00,0xEE,0xFF,0x3F,0x00,0xE6,0xFF,0x3F,0x00,0x02, // 205
+  0x08,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0xE6,0xFF,0x3F,0x00,0xE6,0xFF,0x3F,0x00,0x0C,0x00,0x00,0x00,0x08, // 206
+  0x0C,0x00,0x00,0x00,0x0C,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x0C,0x00,0x00,0x00,0x0C, // 207
+  0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x30,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0xE0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x03,0x0E,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 208
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0xC0,0x01,0x00,0x00,0x80,0x03,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x3C,0x00,0x00,0x08,0x70,0x00,0x00,0x0E,0xE0,0x01,0x00,0x06,0x80,0x03,0x00,0x02,0x00,0x0F,0x00,0x00,0x00,0x1C,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F, // 209
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0xE0,0x00,0x38,0x00,0x62,0x00,0x30,0x00,0x66,0x00,0x30,0x00,0x6E,0x00,0x30,0x00,0x68,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0xE0,0x00,0x38,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x07,0x0F,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 210
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0xE0,0x00,0x38,0x00,0x60,0x00,0x30,0x00,0x68,0x00,0x30,0x00,0x6E,0x00,0x30,0x00,0x66,0x00,0x30,0x00,0x62,0x00,0x30,0x00,0xE0,0x00,0x38,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x07,0x0F,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 211
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0xE0,0x00,0x38,0x00,0x68,0x00,0x30,0x00,0x6C,0x00,0x30,0x00,0x66,0x00,0x30,0x00,0x66,0x00,0x30,0x00,0x6C,0x00,0x30,0x00,0xE8,0x00,0x38,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x07,0x0F,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 212
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xCC,0x00,0x18,0x00,0xEE,0x00,0x38,0x00,0x66,0x00,0x30,0x00,0x66,0x00,0x30,0x00,0x6C,0x00,0x30,0x00,0x6C,0x00,0x30,0x00,0x6E,0x00,0x30,0x00,0xE6,0x00,0x38,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x07,0x0F,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 213
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x01,0x00,0x00,0xFF,0x07,0x00,0x80,0x07,0x0F,0x00,0xC0,0x01,0x1C,0x00,0xC0,0x00,0x18,0x00,0xE0,0x00,0x38,0x00,0x6C,0x00,0x30,0x00,0x6C,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x6C,0x00,0x30,0x00,0xEC,0x00,0x38,0x00,0xC0,0x00,0x18,0x00,0xC0,0x01,0x1C,0x00,0x80,0x07,0x0F,0x00,0x00,0xFF,0x07,0x00,0x00,0xFC,0x01, // 214
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x03,0x00,0x00,0x8E,0x03,0x00,0x00,0xDC,0x01,0x00,0x00,0xF8,0x00,0x00,0x00,0x70,0x00,0x00,0x00,0xF8,0x00,0x00,0x00,0xDC,0x01,0x00,0x00,0x8E,0x03,0x00,0x00,0x06,0x03, // 215
+  0x00,0x00,0x00,0x00,0x00,0xFC,0x21,0x00,0x00,0xFF,0x77,0x00,0x80,0x07,0x3F,0x00,0xC0,0x01,0x1E,0x00,0xC0,0x00,0x1F,0x00,0xE0,0x80,0x3B,0x00,0x60,0xC0,0x31,0x00,0x60,0xE0,0x30,0x00,0x60,0x70,0x30,0x00,0x60,0x38,0x30,0x00,0x60,0x1C,0x30,0x00,0xE0,0x0E,0x38,0x00,0xC0,0x07,0x18,0x00,0xC0,0x03,0x1C,0x00,0xE0,0x07,0x0F,0x00,0x70,0xFF,0x07,0x00,0x20,0xFC,0x01, // 216
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x03,0x00,0xE0,0xFF,0x0F,0x00,0x00,0x00,0x1C,0x00,0x00,0x00,0x38,0x00,0x02,0x00,0x30,0x00,0x06,0x00,0x30,0x00,0x0E,0x00,0x30,0x00,0x08,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x1C,0x00,0xE0,0xFF,0x0F,0x00,0xE0,0xFF,0x03, // 217
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x03,0x00,0xE0,0xFF,0x0F,0x00,0x00,0x00,0x1C,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x30,0x00,0x08,0x00,0x30,0x00,0x0E,0x00,0x30,0x00,0x06,0x00,0x30,0x00,0x02,0x00,0x30,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x1C,0x00,0xE0,0xFF,0x0F,0x00,0xE0,0xFF,0x03, // 218
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x03,0x00,0xE0,0xFF,0x0F,0x00,0x00,0x00,0x1C,0x00,0x00,0x00,0x38,0x00,0x08,0x00,0x30,0x00,0x0C,0x00,0x30,0x00,0x06,0x00,0x30,0x00,0x06,0x00,0x30,0x00,0x0C,0x00,0x30,0x00,0x08,0x00,0x38,0x00,0x00,0x00,0x1C,0x00,0xE0,0xFF,0x0F,0x00,0xE0,0xFF,0x03, // 219
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x03,0x00,0xE0,0xFF,0x0F,0x00,0x00,0x00,0x1C,0x00,0x00,0x00,0x38,0x00,0x0C,0x00,0x30,0x00,0x0C,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x0C,0x00,0x30,0x00,0x0C,0x00,0x38,0x00,0x00,0x00,0x1C,0x00,0xE0,0xFF,0x0F,0x00,0xE0,0xFF,0x03, // 220
+  0x20,0x00,0x00,0x00,0x60,0x00,0x00,0x00,0xC0,0x01,0x00,0x00,0x80,0x03,0x00,0x00,0x00,0x07,0x00,0x00,0x00,0x1E,0x00,0x00,0x00,0x3C,0x00,0x00,0x08,0xF0,0x3F,0x00,0x0E,0xF0,0x3F,0x00,0x06,0x3C,0x00,0x00,0x02,0x1E,0x00,0x00,0x00,0x07,0x00,0x00,0xC0,0x03,0x00,0x00,0xE0,0x01,0x00,0x00,0x60,0x00,0x00,0x00,0x20, // 221
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0x3F,0x00,0xE0,0xFF,0x3F,0x00,0x00,0x03,0x06,0x00,0x00,0x03,0x06,0x00,0x00,0x03,0x06,0x00,0x00,0x03,0x06,0x00,0x00,0x03,0x06,0x00,0x00,0x03,0x06,0x00,0x00,0x03,0x06,0x00,0x00,0x03,0x07,0x00,0x00,0x86,0x03,0x00,0x00,0xFE,0x01,0x00,0x00,0xF8, // 222
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0xFF,0x3F,0x00,0xC0,0xFF,0x3F,0x00,0xC0,0x00,0x00,0x00,0x60,0x00,0x08,0x00,0x60,0x00,0x1C,0x00,0x60,0x00,0x38,0x00,0xE0,0x78,0x30,0x00,0xC0,0x7F,0x30,0x00,0x80,0xC7,0x30,0x00,0x00,0x80,0x39,0x00,0x00,0x80,0x1F,0x00,0x00,0x00,0x0F, // 223
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0x00,0x1C,0x1F,0x00,0x00,0x8C,0x39,0x00,0x20,0x86,0x31,0x00,0x60,0x86,0x31,0x00,0xE0,0xC6,0x30,0x00,0x80,0xC6,0x18,0x00,0x00,0xCE,0x0C,0x00,0x00,0xFC,0x1F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x00,0x20, // 224
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0x00,0x1C,0x1F,0x00,0x00,0x8C,0x39,0x00,0x00,0x86,0x31,0x00,0x80,0x86,0x31,0x00,0xE0,0xC6,0x30,0x00,0x60,0xC6,0x18,0x00,0x20,0xCE,0x0C,0x00,0x00,0xFC,0x1F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x00,0x20, // 225
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0x00,0x1C,0x1F,0x00,0x80,0x8C,0x39,0x00,0xC0,0x86,0x31,0x00,0x60,0x86,0x31,0x00,0x60,0xC6,0x30,0x00,0xC0,0xC6,0x18,0x00,0x80,0xCE,0x0C,0x00,0x00,0xFC,0x1F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x00,0x20, // 226
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0xC0,0x1C,0x1F,0x00,0xE0,0x8C,0x39,0x00,0x60,0x86,0x31,0x00,0x60,0x86,0x31,0x00,0xC0,0xC6,0x30,0x00,0xC0,0xC6,0x18,0x00,0xE0,0xCE,0x0C,0x00,0x60,0xFC,0x1F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x00,0x20, // 227
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0x00,0x1C,0x1F,0x00,0xC0,0x8C,0x39,0x00,0xC0,0x86,0x31,0x00,0x00,0x86,0x31,0x00,0x00,0xC6,0x30,0x00,0xC0,0xC6,0x18,0x00,0xC0,0xCE,0x0C,0x00,0x00,0xFC,0x1F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x00,0x20, // 228
+  0x00,0x00,0x00,0x00,0x00,0x18,0x0E,0x00,0x00,0x1C,0x1F,0x00,0x00,0x8C,0x39,0x00,0x70,0x86,0x31,0x00,0x88,0x86,0x31,0x00,0x88,0xC6,0x30,0x00,0x88,0xC6,0x18,0x00,0x70,0xCE,0x0C,0x00,0x00,0xFC,0x1F,0x00,0x00,0xF8,0x3F,0x00,0x00,0x00,0x20, // 229
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x38,0x00,0x00,0x06,0x30,0x00,0x80,0x06,0x30,0x00,0xE0,0x06,0x30,0x00,0x60,0x0E,0x38,0x00,0x20,0x1C,0x1C,0x00,0x00,0x18,0x0C, // 230
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x38,0x02,0x00,0x06,0x30,0x02,0x00,0x06,0xF0,0x02,0x00,0x06,0xB0,0x03,0x00,0x0E,0x38,0x01,0x00,0x1C,0x1C,0x00,0x00,0x18,0x0C, // 231
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0xDC,0x1C,0x00,0x20,0xCE,0x38,0x00,0x60,0xC6,0x30,0x00,0xE0,0xC6,0x30,0x00,0x80,0xC6,0x30,0x00,0x00,0xCE,0x38,0x00,0x00,0xDC,0x18,0x00,0x00,0xF8,0x0C,0x00,0x00,0xF0,0x04, // 232
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0xDC,0x1C,0x00,0x00,0xCE,0x38,0x00,0x80,0xC6,0x30,0x00,0xE0,0xC6,0x30,0x00,0x60,0xC6,0x30,0x00,0x20,0xCE,0x38,0x00,0x00,0xDC,0x18,0x00,0x00,0xF8,0x0C,0x00,0x00,0xF0,0x04, // 233
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0xDC,0x1C,0x00,0x00,0xCE,0x38,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x30,0x02,0x00,0xC6,0x30,0x02,0x00,0xCE,0xF8,0x02,0x00,0xDC,0x98,0x03,0x00,0xF8,0x0C,0x01,0x00,0xF0,0x04, // 234
+  0x00,0x00,0x00,0x00,0x00,0xE0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0xDC,0x1C,0x00,0xC0,0xCE,0x38,0x00,0xC0,0xC6,0x30,0x00,0x00,0xC6,0x30,0x00,0x00,0xC6,0x30,0x00,0xC0,0xCE,0x38,0x00,0xC0,0xDC,0x18,0x00,0x00,0xF8,0x0C,0x00,0x00,0xF0,0x04, // 235
+  0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x00,0x60,0xFE,0x3F,0x00,0xE0,0xFE,0x3F,0x00,0x80, // 236
+  0x00,0x00,0x00,0x00,0x80,0x00,0x00,0x00,0xE0,0xFE,0x3F,0x00,0x60,0xFE,0x3F,0x00,0x20, // 237
+  0x80,0x00,0x00,0x00,0xC0,0x00,0x00,0x00,0x60,0xFE,0x3F,0x00,0x60,0xFE,0x3F,0x00,0xC0,0x00,0x00,0x00,0x80, // 238
+  0xC0,0x00,0x00,0x00,0xC0,0x00,0x00,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F,0x00,0xC0,0x00,0x00,0x00,0xC0, // 239
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1D,0x1C,0x00,0xA0,0x0F,0x38,0x00,0xA0,0x06,0x30,0x00,0xE0,0x06,0x30,0x00,0xC0,0x06,0x30,0x00,0xC0,0x0F,0x38,0x00,0x20,0x1F,0x1C,0x00,0x00,0xFC,0x0F,0x00,0x00,0xE0,0x07, // 240
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F,0x00,0x00,0x18,0x00,0x00,0x00,0x0C,0x00,0x00,0x80,0x06,0x00,0x00,0xE0,0x06,0x00,0x00,0x60,0x06,0x00,0x00,0x20,0x0E,0x00,0x00,0x00,0xFC,0x3F,0x00,0x00,0xF8,0x3F, // 241
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x20,0x0E,0x38,0x00,0x60,0x06,0x30,0x00,0xE0,0x06,0x30,0x00,0x80,0x06,0x30,0x00,0x00,0x0E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xF0,0x07, // 242
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x38,0x00,0x80,0x06,0x30,0x00,0xE0,0x06,0x30,0x00,0x60,0x06,0x30,0x00,0x20,0x0E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xF0,0x07, // 243
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0x80,0x0E,0x38,0x00,0xC0,0x06,0x30,0x00,0x60,0x06,0x30,0x00,0x60,0x06,0x30,0x00,0xC0,0x0E,0x38,0x00,0x80,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xF0,0x07, // 244
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0xC0,0x1C,0x1C,0x00,0xE0,0x0E,0x38,0x00,0x60,0x06,0x30,0x00,0x60,0x06,0x30,0x00,0xC0,0x06,0x30,0x00,0xC0,0x0E,0x38,0x00,0xE0,0x1C,0x1C,0x00,0x60,0xF8,0x0F,0x00,0x00,0xF0,0x07, // 245
+  0x00,0x00,0x00,0x00,0x00,0xF0,0x07,0x00,0x00,0xF8,0x0F,0x00,0x00,0x1C,0x1C,0x00,0xC0,0x0E,0x38,0x00,0xC0,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0xC0,0x0E,0x38,0x00,0xC0,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xF0,0x07, // 246
+  0x00,0x00,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0xB6,0x01,0x00,0x00,0xB6,0x01,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30, // 247
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xF0,0x67,0x00,0x00,0xF8,0x7F,0x00,0x00,0x1C,0x1C,0x00,0x00,0x0E,0x3F,0x00,0x00,0x86,0x33,0x00,0x00,0xE6,0x31,0x00,0x00,0x76,0x30,0x00,0x00,0x3E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0xFF,0x0F,0x00,0x00,0xF3,0x07, // 248
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x0F,0x00,0x00,0xFE,0x1F,0x00,0x20,0x00,0x38,0x00,0x60,0x00,0x30,0x00,0xE0,0x00,0x30,0x00,0x80,0x00,0x30,0x00,0x00,0x00,0x18,0x00,0x00,0x00,0x0C,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F, // 249
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x0F,0x00,0x00,0xFE,0x1F,0x00,0x00,0x00,0x38,0x00,0x00,0x00,0x30,0x00,0x80,0x00,0x30,0x00,0xE0,0x00,0x30,0x00,0x60,0x00,0x18,0x00,0x20,0x00,0x0C,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F, // 250
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x0F,0x00,0x00,0xFE,0x1F,0x00,0x80,0x00,0x38,0x00,0xC0,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0x60,0x00,0x30,0x00,0xC0,0x00,0x18,0x00,0x80,0x00,0x0C,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F, // 251
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFE,0x0F,0x00,0x00,0xFE,0x1F,0x00,0xC0,0x00,0x38,0x00,0xC0,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0x00,0x00,0x30,0x00,0xC0,0x00,0x18,0x00,0xC0,0x00,0x0C,0x00,0x00,0xFE,0x3F,0x00,0x00,0xFE,0x3F, // 252
+  0x00,0x00,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x7E,0x00,0x06,0x00,0xF0,0x01,0x06,0x00,0x80,0x0F,0x07,0x80,0x00,0xFE,0x03,0xE0,0x00,0xFC,0x00,0x60,0xC0,0x1F,0x00,0x20,0xF8,0x03,0x00,0x00,0x3E,0x00,0x00,0x00,0x06, // 253
+  0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xE0,0xFF,0xFF,0x07,0xE0,0xFF,0xFF,0x07,0x00,0x1C,0x18,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x06,0x30,0x00,0x00,0x0E,0x38,0x00,0x00,0x1C,0x1C,0x00,0x00,0xF8,0x0F,0x00,0x00,0xF0,0x03, // 254
+  0x00,0x00,0x00,0x00,0x00,0x0E,0x00,0x00,0x00,0x7E,0x00,0x06,0xC0,0xF0,0x01,0x06,0xC0,0x80,0x0F,0x07,0x00,0x00,0xFE,0x03,0x00,0x00,0xFC,0x00,0xC0,0xC0,0x1F,0x00,0xC0,0xF8,0x03,0x00,0x00,0x3E,0x00,0x00,0x00,0x06 // 255
+};
+
+char utf8win1250(const uint8_t ch);
+
+const char utf8_win1250_table[] = {
+  0x7A,              // length of table
+
+  // format: 0xaa, 0xbb, 0xcc
+  // 0xaabb (unicode)
+  // 0xcc (win1250)
+  // sorted by unicode to search with bsearch
+  0x00, 0xA0, 0xA0,  // 0xA0 = NO-BREAK SPACE
+  0x00, 0xA4, 0xA4,  // 0xA4 = CURRENCY SIGN
+  0x00, 0xA6, 0xA6,  // 0xA6 = BROKEN BAR
+  0x00, 0xA7, 0xA7,  // 0xA7 = SECTION SIGN
+  0x00, 0xA8, 0xA8,  // 0xA8 = DIAERESIS
+  0x00, 0xA9, 0xA9,  // 0xA9 = COPYRIGHT SIGN
+  0x00, 0xAB, 0xAB,  // 0xAB = LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
+  0x00, 0xAC, 0xAC,  // 0xAC = NOT SIGN
+  0x00, 0xAD, 0xAD,  // 0xAD = SOFT HYPHEN
+  0x00, 0xAE, 0xAE,  // 0xAE = REGISTERED SIGN
+  0x00, 0xB0, 0xB0,  // 0xB0 = DEGREE SIGN
+  0x00, 0xB1, 0xB1,  // 0xB1 = PLUS-MINUS SIGN
+  0x00, 0xB4, 0xB4,  // 0xB4 = ACUTE ACCENT
+  0x00, 0xB5, 0xB5,  // 0xB5 = MICRO SIGN
+  0x00, 0xB6, 0xB6,  // 0xB6 = PILCROW SIGN
+  0x00, 0xB7, 0xB7,  // 0xB7 = MIDDLE DOT
+  0x00, 0xB8, 0xB8,  // 0xB8 = CEDILLA
+  0x00, 0xBB, 0xBB,  // 0xBB = RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
+  0x00, 0xC1, 0xC1,  // 0xC1 = LATIN CAPITAL LETTER A WITH ACUTE
+  0x00, 0xC2, 0xC2,  // 0xC2 = LATIN CAPITAL LETTER A WITH CIRCUMFLEX
+  0x00, 0xC4, 0xC4,  // 0xC4 = LATIN CAPITAL LETTER A WITH DIAERESIS
+  0x00, 0xC7, 0xC7,  // 0xC7 = LATIN CAPITAL LETTER C WITH CEDILLA
+  0x00, 0xC9, 0xC9,  // 0xC9 = LATIN CAPITAL LETTER E WITH ACUTE
+  0x00, 0xCB, 0xCB,  // 0xCB = LATIN CAPITAL LETTER E WITH DIAERESIS
+  0x00, 0xCD, 0xCD,  // 0xCD = LATIN CAPITAL LETTER I WITH ACUTE
+  0x00, 0xCE, 0xCE,  // 0xCE = LATIN CAPITAL LETTER I WITH CIRCUMFLEX
+  0x00, 0xD3, 0xD3,  // 0xD3 = LATIN CAPITAL LETTER O WITH ACUTE
+  0x00, 0xD4, 0xD4,  // 0xD4 = LATIN CAPITAL LETTER O WITH CIRCUMFLEX
+  0x00, 0xD6, 0xD6,  // 0xD6 = LATIN CAPITAL LETTER O WITH DIAERESIS
+  0x00, 0xD7, 0xD7,  // 0xD7 = MULTIPLICATION SIGN
+  0x00, 0xDA, 0xDA,  // 0xDA = LATIN CAPITAL LETTER U WITH ACUTE
+  0x00, 0xDC, 0xDC,  // 0xDC = LATIN CAPITAL LETTER U WITH DIAERESIS
+  0x00, 0xDD, 0xDD,  // 0xDD = LATIN CAPITAL LETTER Y WITH ACUTE
+  0x00, 0xDF, 0xDF,  // 0xDF = LATIN SMALL LETTER SHARP S
+  0x00, 0xE1, 0xE1,  // 0xE1 = LATIN SMALL LETTER A WITH ACUTE
+  0x00, 0xE2, 0xE2,  // 0xE2 = LATIN SMALL LETTER A WITH CIRCUMFLEX
+  0x00, 0xE4, 0xE4,  // 0xE4 = LATIN SMALL LETTER A WITH DIAERESIS
+  0x00, 0xE7, 0xE7,  // 0xE7 = LATIN SMALL LETTER C WITH CEDILLA
+  0x00, 0xE9, 0xE9,  // 0xE9 = LATIN SMALL LETTER E WITH ACUTE
+  0x00, 0xEB, 0xEB,  // 0xEB = LATIN SMALL LETTER E WITH DIAERESIS
+  0x00, 0xED, 0xED,  // 0xED = LATIN SMALL LETTER I WITH ACUTE
+  0x00, 0xEE, 0xEE,  // 0xEE = LATIN SMALL LETTER I WITH CIRCUMFLEX
+  0x00, 0xF3, 0xF3,  // 0xF3 = LATIN SMALL LETTER O WITH ACUTE
+  0x00, 0xF4, 0xF4,  // 0xF4 = LATIN SMALL LETTER O WITH CIRCUMFLEX
+  0x00, 0xF6, 0xF6,  // 0xF6 = LATIN SMALL LETTER O WITH DIAERESIS
+  0x00, 0xF7, 0xF7,  // 0xF7 = DIVISION SIGN
+  0x00, 0xFA, 0xFA,  // 0xFA = LATIN SMALL LETTER U WITH ACUTE
+  0x00, 0xFC, 0xFC,  // 0xFC = LATIN SMALL LETTER U WITH DIAERESIS
+  0x00, 0xFD, 0xFD,  // 0xFD = LATIN SMALL LETTER Y WITH ACUTE
+  0x01, 0x02, 0xC3,  // 0xC3 = LATIN CAPITAL LETTER A WITH BREVE
+  0x01, 0x03, 0xE3,  // 0xE3 = LATIN SMALL LETTER A WITH BREVE
+  0x01, 0x04, 0xA5,  // 0xA5 = LATIN CAPITAL LETTER A WITH OGONEK
+  0x01, 0x05, 0xB9,  // 0xB9 = LATIN SMALL LETTER A WITH OGONEK
+  0x01, 0x06, 0xC6,  // 0xC6 = LATIN CAPITAL LETTER C WITH ACUTE
+  0x01, 0x07, 0xE6,  // 0xE6 = LATIN SMALL LETTER C WITH ACUTE
+  0x01, 0x0C, 0xC8,  // 0xC8 = LATIN CAPITAL LETTER C WITH CARON
+  0x01, 0x0D, 0xE8,  // 0xE8 = LATIN SMALL LETTER C WITH CARON
+  0x01, 0x0E, 0xCF,  // 0xCF = LATIN CAPITAL LETTER D WITH CARON
+  0x01, 0x0F, 0xEF,  // 0xEF = LATIN SMALL LETTER D WITH CARON
+  0x01, 0x10, 0xD0,  // 0xD0 = LATIN CAPITAL LETTER D WITH STROKE
+  0x01, 0x11, 0xF0,  // 0xF0 = LATIN SMALL LETTER D WITH STROKE
+  0x01, 0x18, 0xCA,  // 0xCA = LATIN CAPITAL LETTER E WITH OGONEK
+  0x01, 0x19, 0xEA,  // 0xEA = LATIN SMALL LETTER E WITH OGONEK
+  0x01, 0x1A, 0xCC,  // 0xCC = LATIN CAPITAL LETTER E WITH CARON
+  0x01, 0x1B, 0xEC,  // 0xEC = LATIN SMALL LETTER E WITH CARON
+  0x01, 0x39, 0xC5,  // 0xC5 = LATIN CAPITAL LETTER L WITH ACUTE
+  0x01, 0x3A, 0xE5,  // 0xE5 = LATIN SMALL LETTER L WITH ACUTE
+  0x01, 0x3D, 0xBC,  // 0xBC = LATIN CAPITAL LETTER L WITH CARON
+  0x01, 0x3E, 0xBE,  // 0xBE = LATIN SMALL LETTER L WITH CARON
+  0x01, 0x41, 0xA3,  // 0xA3 = LATIN CAPITAL LETTER L WITH STROKE
+  0x01, 0x42, 0xB3,  // 0xB3 = LATIN SMALL LETTER L WITH STROKE
+  0x01, 0x43, 0xD1,  // 0xD1 = LATIN CAPITAL LETTER N WITH ACUTE
+  0x01, 0x44, 0xF1,  // 0xF1 = LATIN SMALL LETTER N WITH ACUTE
+  0x01, 0x47, 0xD2,  // 0xD2 = LATIN CAPITAL LETTER N WITH CARON
+  0x01, 0x48, 0xF2,  // 0xF2 = LATIN SMALL LETTER N WITH CARON
+  0x01, 0x50, 0xD5,  // 0xD5 = LATIN CAPITAL LETTER O WITH DOUBLE ACUTE
+  0x01, 0x51, 0xF5,  // 0xF5 = LATIN SMALL LETTER O WITH DOUBLE ACUTE
+  0x01, 0x54, 0xC0,  // 0xC0 = LATIN CAPITAL LETTER R WITH ACUTE
+  0x01, 0x55, 0xE0,  // 0xE0 = LATIN SMALL LETTER R WITH ACUTE
+  0x01, 0x58, 0xD8,  // 0xD8 = LATIN CAPITAL LETTER R WITH CARON
+  0x01, 0x59, 0xF8,  // 0xF8 = LATIN SMALL LETTER R WITH CARON
+  0x01, 0x5A, 0x8C,  // 0x8C = LATIN CAPITAL LETTER S WITH ACUTE
+  0x01, 0x5B, 0x9C,  // 0x9C = LATIN SMALL LETTER S WITH ACUTE
+  0x01, 0x5E, 0xAA,  // 0xAA = LATIN CAPITAL LETTER S WITH CEDILLA
+  0x01, 0x5F, 0xBA,  // 0xBA = LATIN SMALL LETTER S WITH CEDILLA
+  0x01, 0x60, 0x8A,  // 0x8A = LATIN CAPITAL LETTER S WITH CARON
+  0x01, 0x61, 0x9A,  // 0x9A = LATIN SMALL LETTER S WITH CARON
+  0x01, 0x62, 0xDE,  // 0xDE = LATIN CAPITAL LETTER T WITH CEDILLA
+  0x01, 0x63, 0xFE,  // 0xFE = LATIN SMALL LETTER T WITH CEDILLA
+  0x01, 0x64, 0x8D,  // 0x8D = LATIN CAPITAL LETTER T WITH CARON
+  0x01, 0x65, 0x9D,  // 0x9D = LATIN SMALL LETTER T WITH CARON
+  0x01, 0x6E, 0xD9,  // 0xD9 = LATIN CAPITAL LETTER U WITH RING ABOVE
+  0x01, 0x6F, 0xF9,  // 0xF9 = LATIN SMALL LETTER U WITH RING ABOVE
+  0x01, 0x70, 0xDB,  // 0xDB = LATIN CAPITAL LETTER U WITH DOUBLE ACUTE
+  0x01, 0x71, 0xFB,  // 0xFB = LATIN SMALL LETTER U WITH DOUBLE ACUTE
+  0x01, 0x79, 0x8F,  // 0x8F = LATIN CAPITAL LETTER Z WITH ACUTE
+  0x01, 0x7A, 0x9F,  // 0x9F = LATIN SMALL LETTER Z WITH ACUTE
+  0x01, 0x7B, 0xAF,  // 0xAF = LATIN CAPITAL LETTER Z WITH DOT ABOVE
+  0x01, 0x7C, 0xBF,  // 0xBF = LATIN SMALL LETTER Z WITH DOT ABOVE
+  0x01, 0x7D, 0x8E,  // 0x8E = LATIN CAPITAL LETTER Z WITH CARON
+  0x01, 0x7E, 0x9E,  // 0x9E = LATIN SMALL LETTER Z WITH CARON
+  0x02, 0xC7, 0xA1,  // 0xA1 = CARON
+  0x02, 0xD8, 0xA2,  // 0xA2 = BREVE
+  0x02, 0xD9, 0xFF,  // 0xFF = DOT ABOVE
+  0x02, 0xDB, 0xB2,  // 0xB2 = OGONEK
+  0x02, 0xDD, 0xBD,  // 0xBD = DOUBLE ACUTE ACCENT
+  0x20, 0x13, 0x96,  // 0x96 = EN DASH
+  0x20, 0x14, 0x97,  // 0x97 = EM DASH
+  0x20, 0x18, 0x91,  // 0x91 = LEFT SINGLE QUOTATION MARK
+  0x20, 0x19, 0x92,  // 0x92 = RIGHT SINGLE QUOTATION MARK
+  0x20, 0x1A, 0x82,  // 0x82 = SINGLE LOW-9 QUOTATION MARK
+  0x20, 0x1C, 0x93,  // 0x93 = LEFT DOUBLE QUOTATION MARK
+  0x20, 0x1D, 0x94,  // 0x94 = RIGHT DOUBLE QUOTATION MARK
+  0x20, 0x1E, 0x84,  // 0x84 = DOUBLE LOW-9 QUOTATION MARK
+  0x20, 0x20, 0x86,  // 0x86 = DAGGER
+  0x20, 0x21, 0x87,  // 0x87 = DOUBLE DAGGER
+  0x20, 0x22, 0x95,  // 0x95 = BULLET
+  0x20, 0x26, 0x85,  // 0x85 = HORIZONTAL ELLIPSIS
+  0x20, 0x30, 0x89,  // 0x89 = PER MILLE SIGN
+  0x20, 0x39, 0x8B,  // 0x8B = SINGLE LEFT-POINTING ANGLE QUOTATION MARK
+  0x20, 0x3A, 0x9B,  // 0x9B = SINGLE RIGHT-POINTING ANGLE QUOTATION MARK
+  0x20, 0xAC, 0x80,  // 0x80 = EURO SIGN
+  0x21, 0x22, 0x99   // 0x99 = TRADE MARK SIGN
+};
 #endif
+
 #endif  // SuplaOled_H
